### 2023-11-27 #6094 @editorialbot
[PRE REVIEW]: ABSESpy: An agent-based modeling framework for social-ecological systems
**Submitting author:** @SongshGeo (<a href="http://orcid.org/0000-0002-8112-8780">Shuang Song</a>)
**Repository:** https://github.com/ABSESpy/ABSESpy
**Branch with paper.md** (empty if default branch): dev
**Version:** v0.3.0
