### 2023-11-28 6b94d1d
build(projects): :building_construction: remove pint and prettytable dependencies
