+++ pyproject.toml
@@ -39,8 +39,6 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = ">=3.9,<3.12"
-prettytable = "^3.6.0"
-pint = "^0.20.1"
 netcdf4 = "^1.6.2"
 hydra-core = "^1.3.2"
 mesa-geo = "^0.6.0"
