### 2023-11-28 da7b50e
build(projects): :arrow_up: cancel pin of deps and upgrade
