+++ requirements.txt
@@ -31,7 +31,7 @@ filelock==3.13.1 ; python_version >= "3.9" and python_version < "3.12"
 fiona==1.9.5 ; python_version >= "3.9" and python_version < "3.12"
 folium==0.15.0 ; python_version >= "3.9" and python_version < "3.12"
 fonttools==4.45.1 ; python_version >= "3.9" and python_version < "3.12"
-geopandas==0.12.2 ; python_version >= "3.9" and python_version < "3.12"
+geopandas==0.14.1 ; python_version >= "3.9" and python_version < "3.12"
 h11==0.14.0 ; python_version >= "3.9" and python_version < "3.12"
 humanize==4.9.0 ; python_version >= "3.9" and python_version < "3.12"
 hydra-core==1.3.2 ; python_version >= "3.9" and python_version < "3.12"
@@ -66,10 +66,10 @@ nbformat==5.9.2 ; python_version >= "3.9" and python_version < "3.12"
 nest-asyncio==1.5.8 ; python_version >= "3.9" and python_version < "3.12"
 netcdf4==1.6.5 ; python_version >= "3.9" and python_version < "3.12"
 networkx==3.2.1 ; python_version >= "3.9" and python_version < "3.12"
-numpy==1.26.2 ; python_version < "3.12" and python_version >= "3.9"
+numpy==1.26.2 ; python_version >= "3.9" and python_version < "3.12"
 omegaconf==2.3.0 ; python_version >= "3.9" and python_version < "3.12"
 packaging==23.2 ; python_version >= "3.9" and python_version < "3.12"
-pandas==1.5.3 ; python_version >= "3.9" and python_version < "3.12"
+pandas==2.1.3 ; python_version >= "3.9" and python_version < "3.12"
 parso==0.8.3 ; python_version >= "3.9" and python_version < "3.12"
 pendulum==2.1.2 ; python_version >= "3.9" and python_version < "3.12"
 pexpect==4.9.0 ; python_version >= "3.9" and python_version < "3.12" and sys_platform != "win32"
@@ -116,6 +116,7 @@ tqdm==4.66.1 ; python_version >= "3.9" and python_version < "3.12"
 traitlets==5.14.0 ; python_version >= "3.9" and python_version < "3.12"
 types-python-dateutil==2.8.19.14 ; python_version >= "3.9" and python_version < "3.12"
 typing-extensions==4.8.0 ; python_version >= "3.9" and python_version < "3.12"
+tzdata==2023.3 ; python_version >= "3.9" and python_version < "3.12"
 urllib3==2.1.0 ; python_version >= "3.9" and python_version < "3.12"
 uvicorn==0.24.0.post1 ; python_version >= "3.9" and python_version < "3.12"
 watchdog==3.0.0 ; python_version >= "3.9" and python_version < "3.12"
