+++ pyproject.toml
@@ -39,18 +39,17 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = ">=3.9,<3.12"
-netcdf4 = "^1.6.2"
-hydra-core = "^1.3.2"
+netcdf4 = ">=1.6"
+hydra-core = "~1.3"
 mesa-geo = "^0.6.0"
-xarray = "^2023.8.0"
-fiona = "^1.9.4.post1"
-loguru = "^0.7.2"
-rioxarray = "^0.13.3"
-pendulum = "^2.1.2"
-rich = "^13.6.0"
-numpy = "^1.24.1"
-pandas = "^1.5.3"
-geopandas = "^0.12.2"
+xarray = "~2023"
+fiona = ">1.8"
+loguru = "~0.7"
+rioxarray = "~0.13"
+pendulum = "~2"
+numpy = ">=1.24"
+pandas = ">=1.4,<3"
+geopandas = "~0"
 
 [tool.poetry.group.dev.dependencies]
 pytest-clarity = "^1.0.1"
