### 2023-11-29 fd23432
chore(projects): :bookmark: update v0.3.2, introduced a random module
