### 2023-11-29 07a9f66
fix(modules): :ambulance: fixing ActorsList.random.choice triggered error
