+++ abses/sequences.py
@@ -69,11 +69,11 @@ class ActorsList(list):
     def __getattr__(self, name: str) -> np.ndarray:
         """Return callable list of attributes"""
         # Private variables are looked up normally
+        if name == "random":
+            return getattr(super(), name)
         if name[0] == "_":
             return getattr(super(), name)
         if name in self.__dir__():
-            return getattr(super(), name)
-        if name == "random":
             return getattr(super(), name)
         return ActorsList.array(self, name)
 
