### 2023-11-29 6d134d7
fix(modules): :ambulance: fixing ActorsList.random.choice triggered error
