### 2023-12-03 75897d3
Added some tests for the datacollection module
