+++ tests/test_datacollection.py
@@ -1,0 +1,314 @@
+#!/usr/bin/env python 3.11.0
+# -*-coding:utf-8 -*-
+# @Author  : Shuang (Twist) Song
+# @Contact   : SongshGeo@gmail.com
+# GitHub   : https://github.com/SongshGeo
+# Website: https://cv.songshgeo.com/
+
+import numpy as np
+import pytest
+
+from abses import Actor, MainModel
+from abses.datacollection import DataCollector
+
+rng = np.random.default_rng(seed=42)
+
+
+class MockAgent(Actor):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.val = 0
+        self.val2 = 2
+
+
+class anotherMockAgent(MockAgent):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.val3 = 3
+
+
+class MockModel(MainModel):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.model_var = 100
+        self.n = 10
+
+        self.agents.create(MockAgent, self.n - 3)
+        self.agents.create(anotherMockAgent, 3)
+
+        self.datacollector = DataCollector(
+            model=self,
+            model_reporters={
+                "const": "model_var",
+                "pop_ratio": lambda m: len(m.agents["MockAgent"])
+                / len(m.agents["anotherMockAgent"]),
+            },
+            agent_reporters={
+                "const": "val2",
+            },
+        )
+
+    def step(self):
+        if self.time.tick == 5:
+            self.agents.remove(self.actors[5])
+
+        self.datacollector.collect()
+
+
+class Test_DataCollector:
+    def setup_method(self):
+        self.model = MockModel(parameters={"time": {"end": 10}})
+
+        self.model.run_model()
+
+    def test_model_vars(self):
+        datacollector = self.model.datacollector
+        assert "const" in datacollector.model_vars
+        assert "pop_ratio" in datacollector.model_vars
+        ratio = 2.3333333333333335  # Implied ratio at beginning
+        assert datacollector.model_vars["pop_ratio"][1] == ratio
+        ratio = 2.0  # Implied ratio at end
+        assert datacollector.model_vars["pop_ratio"][-1] == ratio
+
+    def test_agent_records(self):
+        pass
+
+    def test_table_rows(self):
+        pass
+
+
+"""
+""
+Test the DataCollector
+""
+import unittest
+
+from mesa import Agent, Model
+from mesa.time import BaseScheduler
+
+
+class MockAgent(Agent):
+    ""
+    Minimalistic agent for testing purposes.
+    ""
+
+    def __init__(self, unique_id, model, val=0):
+        super().__init__(unique_id, model)
+        self.val = val
+        self.val2 = val
+
+    def step(self):
+        ""
+        Increment vals by 1.
+        ""
+        self.val += 1
+        self.val2 += 1
+
+    def double_val(self):
+        return self.val * 2
+
+    def write_final_values(self):
+        ""
+        Write the final value to the appropriate table.
+        ""
+        row = {"agent_id": self.unique_id, "final_value": self.val}
+        self.model.datacollector.add_table_row("Final_Values", row)
+
+
+def agent_function_with_params(agent, multiplier, offset):
+    return (agent.val * multiplier) + offset
+
+
+class DifferentMockAgent(MockAgent):
+    # We define a different MockAgent to test for attributes that are present
+    # only in 1 type of agent, but not the other.
+    def __init__(self, unique_id, model, val=0):
+        super().__init__(unique_id, model, val=val)
+        self.val3 = val + 42
+
+
+class MockModel(Model):
+    ""
[…]
