### 2023-12-04 3ea59da
Added tests for the datacollection module (improved)
