+++ tests/test_datacollection.py
@@ -75,240 +75,3 @@ class Test_DataCollector:
 
     def test_table_rows(self):
         pass
-
-
-"""
-""
-Test the DataCollector
-""
-import unittest
-
-from mesa import Agent, Model
-from mesa.time import BaseScheduler
-
-
-class MockAgent(Agent):
-    ""
-    Minimalistic agent for testing purposes.
-    ""
-
-    def __init__(self, unique_id, model, val=0):
-        super().__init__(unique_id, model)
-        self.val = val
-        self.val2 = val
-
-    def step(self):
-        ""
-        Increment vals by 1.
-        ""
-        self.val += 1
-        self.val2 += 1
-
-    def double_val(self):
-        return self.val * 2
-
-    def write_final_values(self):
-        ""
-        Write the final value to the appropriate table.
-        ""
-        row = {"agent_id": self.unique_id, "final_value": self.val}
-        self.model.datacollector.add_table_row("Final_Values", row)
-
-
-def agent_function_with_params(agent, multiplier, offset):
-    return (agent.val * multiplier) + offset
-
-
-class DifferentMockAgent(MockAgent):
-    # We define a different MockAgent to test for attributes that are present
-    # only in 1 type of agent, but not the other.
-    def __init__(self, unique_id, model, val=0):
-        super().__init__(unique_id, model, val=val)
-        self.val3 = val + 42
-
-
-class MockModel(Model):
-    ""
-    Minimalistic model for testing purposes.
-    ""
-
-    schedule = BaseScheduler(None)
-
-    def __init__(self):
-        self.schedule = BaseScheduler(self)
-        self.model_val = 100
-
-        self.n = 10
-        for i in range(self.n):
-            self.schedule.add(MockAgent(i, self, val=i))
-        self.initialize_data_collector(
-            model_reporters={
-                "total_agents": lambda m: m.schedule.get_agent_count(),
-                "model_value": "model_val",
-                "model_calc": self.schedule.get_agent_count,
-                "model_calc_comp": [self.test_model_calc_comp, [3, 4]],
-                "model_calc_fail": [self.test_model_calc_comp, [12, 0]],
-            },
-            agent_reporters={
-                "value": lambda a: a.val,
-                "value2": "val2",
-                "double_value": MockAgent.double_val,
-                "value_with_params": [agent_function_with_params, [2, 3]],
-            },
-            tables={"Final_Values": ["agent_id", "final_value"]},
-        )
-
-    def test_model_calc_comp(self, input1, input2):
-        if input2 > 0:
-            return (self.model_val * input1) / input2
-        else:
-            assert ValueError
-            return None
-
-    def step(self):
-        self.schedule.step()
-        self.datacollector.collect(self)
-
-
-class TestDataCollector(unittest.TestCase):
-    def setUp(self):
-        ""
-        Create the model and run it a set number of steps.
-        ""
-        self.model = MockModel()
-        for i in range(7):
-            if i == 4:
-                self.model.schedule.remove(self.model.schedule._agents[3])
-            self.model.step()
-
-        # Write to table:
-        for agent in self.model.schedule.agents:
-            agent.write_final_values()
-
-    def step_assertion(self, model_var):
-        for element in model_var:
-            if model_var.index(element) < 4:
-                assert element == 10
-            else:
-                assert element == 9
-
-    def test_model_vars(self):
-        ""
-        Test model-level variable collection.
-        ""
-        data_collector = self.model.datacollector
-        assert "total_agents" in data_collector.model_vars
-        assert "model_value" in data_collector.model_vars
-        assert "model_calc" in data_collector.model_vars
-        assert "model_calc_comp" in data_collector.model_vars
-        assert "model_calc_fail" in data_collector.model_vars
-        length = 8
[…]
