### 2023-12-05 be5946d
feat(modules): :sparkles: random choice
