### 2023-12-05 63370a8
test(projects): :white_check_mark: tests all passed after random.choice
