+++ abses/sequences.py
@@ -165,50 +165,6 @@ class ActorsList(list):
         ids = make_list(ids)
         return self.select([agent.id in ids for agent in self])
 
-    def random_choose(
-        self,
-        size: int = 1,
-        prob: Optional[Iterable[float]] = None,
-        replace: bool = True,
-        as_list: bool = False,
-    ) -> Union[Actor, Self]:
-        """Randomly choose one or more actors from the current self object.
-
-        Parameters:
-            size:
-                The number of actors to choose. Defaults to 1.
-            prob:
-                A list of probabilities for each actor to be chosen.
-                If None, all actors have equal probability. Defaults to None.
-            replace:
-                Whether to sample with replacement. Defaults to True.
-            as_list:
-                Whether to return the result as a list of actors. Defaults to False.
-
-        Returns:
-            An Actor or an ActorList of multiple actors.
-
-        Notes:
-            Given the parameter set size=1 and as_list=False, a single Actor object is returned.
-            Given the parameter set size>1 and as_list=False, a Self (ActorsList) object is returned.
-
-        Raises:
-            ValueError:
-                If size is not a positive integer.
-        """
-        # TODO refactor this to `self.random.choice`
-        logger.warning(
-            "Deprecated Warning: In the next version, use `ActorsList.random.choice` instead of `ActorsList.random_choose`."
-        )
-        chosen = norm_choice(self, p=prob, size=size, replace=replace)
-        if as_list:
-            return ActorsList(self._model, objs=chosen)
-        if size == 1:
-            return chosen[0]
-        if size > 1:
-            return ActorsList(self._model, chosen)
-        raise ValueError(f"Invalid size {size}.")
-
     def better(
         self, metric: str, than: Optional[Union[Number, Actor]] = None
     ) -> Self:
