### 2023-12-05 1fb1c8e
refactor(modules): :fire: remove unuseful norm_choose func
