### 2023-12-05 7a2e3ce
docs(docs): :memo: update api docs for the random module
