### 2023-12-05 dfa15c4
fix(modules): :ambulance: default getting array of ActorsList is deprecated
