### 2023-12-05 aad6cc4
Added more tests for datacollection module and corrected some tests to account for missing default of random seed
