### 2023-12-19 cb25239
Added some new logging capabilities. First round
