### 2024-01-07 04a169d
Added initial contents to movement notebook
