### 2024-01-10 3167c36
build: :arrow_up: using mkdocs material insider
