+++ requirements.txt
@@ -1,10 +1,10 @@
 affine==2.4.0 ; python_version >= "3.9" and python_version < "3.12"
 antlr4-python3-runtime==4.9.3 ; python_version >= "3.9" and python_version < "3.12"
-anyio==4.1.0 ; python_version >= "3.9" and python_version < "3.12"
+anyio==4.2.0 ; python_version >= "3.9" and python_version < "3.12"
 appnope==0.1.3 ; python_version >= "3.9" and python_version < "3.12" and platform_system == "Darwin"
 arrow==1.3.0 ; python_version >= "3.9" and python_version < "3.12"
 asttokens==2.4.1 ; python_version >= "3.9" and python_version < "3.12"
-attrs==23.1.0 ; python_version >= "3.9" and python_version < "3.12"
+attrs==23.2.0 ; python_version >= "3.9" and python_version < "3.12"
 beautifulsoup4==4.12.2 ; python_version >= "3.9" and python_version < "3.12"
 binaryornot==0.4.4 ; python_version >= "3.9" and python_version < "3.12"
 branca==0.7.0 ; python_version >= "3.9" and python_version < "3.12"
@@ -18,7 +18,7 @@ click-plugins==1.1.1 ; python_version >= "3.9" and python_version < "3.12"
 click==8.1.7 ; python_version >= "3.9" and python_version < "3.12"
 cligj==0.7.2 ; python_version >= "3.9" and python_version < "3.12"
 colorama==0.4.6 ; python_version >= "3.9" and python_version < "3.12" and sys_platform == "win32" or python_version >= "3.9" and python_version < "3.12" and platform_system == "Windows"
-comm==0.2.0 ; python_version >= "3.9" and python_version < "3.12"
+comm==0.2.1 ; python_version >= "3.9" and python_version < "3.12"
 contourpy==1.2.0 ; python_version >= "3.9" and python_version < "3.12"
 cookiecutter==2.5.0 ; python_version >= "3.9" and python_version < "3.12"
 cycler==0.12.1 ; python_version >= "3.9" and python_version < "3.12"
@@ -26,29 +26,29 @@ debugpy==1.8.0 ; python_version >= "3.9" and python_version < "3.12"
 decorator==5.1.1 ; python_version >= "3.9" and python_version < "3.12"
 exceptiongroup==1.2.0 ; python_version >= "3.9" and python_version < "3.11"
 executing==2.0.1 ; python_version >= "3.9" and python_version < "3.12"
-fastjsonschema==2.19.0 ; python_version >= "3.9" and python_version < "3.12"
+fastjsonschema==2.19.1 ; python_version >= "3.9" and python_version < "3.12"
 filelock==3.13.1 ; python_version >= "3.9" and python_version < "3.12"
 fiona==1.9.5 ; python_version >= "3.9" and python_version < "3.12"
-folium==0.15.0 ; python_version >= "3.9" and python_version < "3.12"
-fonttools==4.45.1 ; python_version >= "3.9" and python_version < "3.12"
-geopandas==0.14.1 ; python_version >= "3.9" and python_version < "3.12"
+folium==0.15.1 ; python_version >= "3.9" and python_version < "3.12"
+fonttools==4.47.0 ; python_version >= "3.9" and python_version < "3.12"
+geopandas==0.14.2 ; python_version >= "3.9" and python_version < "3.12"
 h11==0.14.0 ; python_version >= "3.9" and python_version < "3.12"
 humanize==4.9.0 ; python_version >= "3.9" and python_version < "3.12"
 hydra-core==1.3.2 ; python_version >= "3.9" and python_version < "3.12"
 idna==3.6 ; python_version >= "3.9" and python_version < "3.12"
[…]
