+++ docs/wiki/definition_ABM.md
@@ -2,10 +2,6 @@
 
 source:: https://gistbok.ucgis.org/bok-topics/agent-based-modeling
 
-Agent-based models (ABMs) in the context of GIScience are dynamic models that simulate how individuals make decisions based on their interactions with others and the space around them. The term “agent-based” has typically been used in the social sciences when such models are developed to represent individual people, while the term “individual-based” was adopted by ecologists and biologists when developing such models to emulate the behaviors of other organisms such as birds, fish, and mammals (hereafter we refer to all such models as “agent-based”). A distinctive feature of ABMs in GIScience and related disciplines is that the models are inherently spatial, as agents exist at some location in space at every moment in time, and can obtain information from their location and the locations around them to make decisions about what to do or where to move.
+Agent-based models (ABMs) in the context of GIScience are dynamic models that simulate how individuals make decisions based on their interactions with others and the space around them. The term “agent-based” has typically been used in the social sciences when such models are developed to represent individual people, while the term ``individual-based`` was adopted by ecologists and biologists when developing such models to emulate the behaviors of other organisms such as birds, fish, and mammals (hereafter we refer to all such models as “agent-based”). A distinctive feature of ABMs in GIScience and related disciplines is that the models are inherently spatial, as agents exist at some location in space at every moment in time, and can obtain information from their location and the locations around them to make decisions about what to do or where to move.
 
 在地理信息科学（GIScience）的背景下，基于代理的模型（ABMs）是动态模型，模拟个体如何根据与他人和周围空间的互动来做出决策。当这样的模型被开发用来表示单个人时，“基于代理”的术语通常在社会科学中使用，而生态学家和生物学家在开发此类模型以仿真其他生物（如鸟类、鱼类和哺乳动物）的行为时，则采用了“基于个体”的术语（以下我们将所有这些模型统称为“基于代理”）。 GIScience及相关领域中ABMs 的一个显著特点是，这些模型本质上具有空间性，因为代理在每一刻都存在于某个空间位置，并且可以从其位置和周围位置获取信息以决定该做什么或移动到哪里。
-
-ABMs have been developed to simulate the decisions people make as they walk through crowds in urban spaces in order to reach a specific destination (Torrens 2012), and the decisions that current and potential home owners make regarding where to move based on the social and physical characteristics of neighborhoods (Parker and Filatova 2008). Similarly, non-human agent models have been developed to represent individual elk migrating through a landscape and the decisions they make in response to food availability, topography, and the overall behavior of their herd (Bennett and Tang 2006). The power of this type of modeling is that the collection of decisions made by all the individual agents produces an emergent spatial pattern over time. In doing so, ABMs can help us learn how crowds move through urban areas at different times in the day, how locational preference creates income and racial segregation in cities, or how food availability and topography influence animal herd migratory patterns. In short, ABMs help us link individual decisions made at the local level to emergent patterns at the level of the overall system.
-
-已经开发了ABMs 来模拟人们穿过城市空间中的人群以达到特定目标所做出的决策 (Torrens 2012)，以及现有房主和潜在房主根据邻居社区的社会和物理特征对搬迁地点进行选择所作出的决策 (Parker and Filatova 2008)。同样地，也开发了非人类代理模型来表示单个麋鹿穿越景观以及他们对食物供应、地形和群体整体行为的反应所做出的决策 (Bennett and Tang 2006)。这种建模方式的强大之处在于，所有单独代理做出的决策集合会随着时间产生一种新兴空间模式。因此，ABMs 可以帮助我们了解人群如何在一天中不同时间通过城市区域，如何定位偏好导致城市中收入和种族隔离，或者食物供应和地形如何影响动物群迁移模式。简而言之，ABMs 帮助我们将局部层面上作出的个别决策与整个系统层面上的新兴模式联系起来。
