+++ includes/abbreviations.md
@@ -1,0 +1,5 @@
+*[ABM]: Agent-based model
+*[SES]: Socio-ecological system
+
+[python]: https://www.python.org/
+[wiki pages]: ../wiki/wiki.md
