+++ docs/index.md
@@ -1,56 +1,57 @@
 ---
-title: index
-authors: SongshGeo
-long_name: Agent-Based Social-ecological systems Modelling Framework in Python
-name: ABSESpy Documentation
-state: review
-banner_icon: 💻
-banner_y: 0.52
-SES: social-ecological systems
-github: https://github.com/SongshGeo
-email: songshgeo@gmail.com
-website: https://cv.songshgeo.com
-glightbox: false
+title: "ABSESpy"
+description: "Agent-Based Social-ecological systems Modelling Framework in Python"
+# status: new
 ---
 ![ABSES_banner](https://songshgeo-picgo-1302043007.cos.ap-beijing.myqcloud.com/uPic/CleanShot%202023-10-19%20at%2019.08.12@2x.png)
 
 <!-- Language: [English Readme](#) | [简体中文](README_ch) -->
 
-**Date**: October 19, 2023, **Version**: 0.2.0.alpha
+**Date**: January. 08, 2024, **Version**: 0.3.5
 
-**Useful links**: [Install](home/install.md) | [Source Repository](https://github.com/ABSESpy/ABSESpy) | [Issues & Ideas](https://github.com/ABSESpy/ABSESpy/issues) | [Q&A Support](https://github.com/ABSESpy/ABSESpy/discussions) | [Mailing List](https://groups.google.com/g/absespy)
+**Useful links**: [Install](home/Installation.md) | [Source Repository](https://github.com/ABSESpy/ABSESpy) | [Issues & Ideas](https://github.com/ABSESpy/ABSESpy/issues) | [Q&A Support](https://github.com/ABSESpy/ABSESpy/discussions)
 
-`ABSESpy` is an Agent-Based modeling (ABM) framework that makes modeling artificial **[Social-ecological systems](https://absespy.github.io/ABSESpy/docs/about/)** (SES) easier.
+`ABSESpy` is an agent-based framework that makes modeling social-ecological systems easier.
 
-<div style="display: flex; justify-content: center;">
-<table>
-    <tr>
-        <td style="text-align: center;">
-            <a href="https://absespy.github.io/ABSESpy/wiki/paper/">
-                <img src="https://songshgeo-picgo-1302043007.cos.ap-beijing.myqcloud.com/uPic/D6J1Gj.png" alt="Step 1" width="300"/>
-            </a>
-            <p>Pure beginner of ABM or SES</p>
-        </td>
-        <td style="text-align: center;">
-            <a href="https://absespy.github.io/ABSESpy/home/install/">
-                <img src="https://songshgeo-picgo-1302043007.cos.ap-beijing.myqcloud.com/uPic/9324755_wiki_wikipedia_book_directory_ebook_icon.png" alt="Step 2" width="300"/>
-            </a>
-            <p>Know basic ideas but don't know how to start</p>
-        </td>
-    </tr>
-    <tr>
-        <td style="text-align: center;">
-            <a href="https://absespy.github.io/ABSESpy/api/model/">
-                <img src="https://songshgeo-picgo-1302043007.cos.ap-beijing.myqcloud.com/uPic/1238311_div_div%20coding_html_html%20coding_source%20code_icon.png" alt="Step 3" width="300"/>
-            </a>
-            <p>Source code and API documentation</p>
-        </td>
-        <td style="text-align: center;">
-            <a href="https://absespy.github.io/ABSESpy/tutorial/lessons/organize_model_structure/">
-                <img src="https://songshgeo-picgo-1302043007.cos.ap-beijing.myqcloud.com/uPic/942329_computer_internet_laptop_learn_teach_icon.png" alt="Step 4" width="300"/>
-            </a>
-            <p>Check out the tutorials</p>
[…]
