+++ abbreviations.md
@@ -1,7 +1,0 @@
-# Abbreviations
-
-*[SES]: Social-ecological System
-*[SESs]: Social-ecological Systems
-*[ABM]: Agent-based Model
-*[IAD]: Institutions Analysis and Development
-*[CAS]: Complex Adaptive Systems
