+++ docs/tutorial/advanced/add_modules.md
@@ -1,3 +1,0 @@
-
-> [!INFO]In Progress
-> This document is a work in progress if you see any errors, or exclusions or have any problems, please [get in touch with us](https://github.com/absespy/ABSESpy/issues).

[…]
