### 2024-01-11 12b6ea5
chore(docs): :coffin: remove died docs
