+++ docs/features/features.md
@@ -1,24 +1,0 @@
----
-title: features
-authors: SongshGeo
-date: 2023-01-10
-long_name: Agent-Based Social-ecological systems Modelling Framework in Python
-name: AB-SESpy
-state: open
-banner_icon: 💻
-banner: "https://songshgeo-picgo-1302043007.cos.ap-beijing.myqcloud.com/uPic/abses_github_repo.svg"
-banner_y: 0.52
-SES: social-ecological systems
-github: https://github.com/SongshGeo
-email: songshgeo@gmail.com
-website: https://cv.songshgeo.com
----
-
-> [!INFO]In Progress
-> This document is a work in progress if you see any errors, or exclusions or have any problems, please [get in touch with us](https://github.com/absespy/ABSESpy/issues).
-
-Agent-based model (ABM) is essential for social-ecological systems (SES) research. `ABSESpy` is designed for modeling **couples humans and nature systems** by:
-
-- [Architectural Elegance for Modular Socio-Ecological Systems Modeling](architectural_elegance.md).
-- [Mastering Time in SES Modeling with Real-world Precision and Dynamic Updates](time_control.md).
-- [Unveiling Human Decision Dynamics in SES Modeling by an Advanced Human Behavior Simulating Framework.](MoHuB.md)
