+++ includes/abbreviations.md
@@ -1,5 +1,6 @@
 *[ABM]: Agent-based model
 *[SES]: Socio-ecological system
+*[CAS]: Complex Adaptive System
 
 [python]: https://www.python.org/
 [wiki pages]: ../wiki/wiki.md
