+++ docs/home/contribution.md
@@ -1,0 +1,18 @@
+# :material-human-handsup: Contributions are welcomed
+
+## Ask questions
+
+- Ask usage questions ("How to do?") on [GitHub Discussions].
+- For less well-defined questions or ideas or to announce other projects of interest to users, use the [mailing list].
+
+## Report bugs & Code contributions
+
+- Report bugs, suggest features or view the source code on [GitHub issues].
+- If you are brand new to _`ABSESpy`_ or open-source development, we recommend going through the [GitHub issues] tab to find issues that interest you.
+<!-- - Once you’ve found an exciting issue, you can return here to get your development environment set up. -->
+
+
+<!-- Links -->
+  [mailing list]: https://groups.google.com/g/absespy
+  [GitHub Discussions]: https://github.com/absespy/ABSESpy/discussions
+  [GitHub issues]: https://github.com/absespy/ABSESpy/issues
