+++ docs/examples/water_quota/negotiation.md
@@ -1,15 +1,0 @@
----
-title: negotiation
-authors: SongshGeo
-date: 2023-01-10
-long_name: Agent-Based Social-ecological systems Modelling Framework in Python
-name: AB-SESpy
-state: open
-banner_icon: 💻
-banner: https://songshgeo-picgo-1302043007.cos.ap-beijing.myqcloud.com/uPic/abses_github_repo.svg
-banner_y: 0.52
-SES: social-ecological systems
-github: https://github.com/SongshGeo
-email: songshgeo@gmail.com
-website: https://cv.songshgeo.com
----
