### 2024-01-11 c0767a5
docs(docs): :memo: update dependencies' doc
