+++ docs/home/dependencies.md
@@ -3,10 +3,13 @@
 
 | Package       | Version           | Purpose                                               |
 |---------------|-------------------|-------------------------------------------------------|
-| python        | >=3.9,<4          | Core programming language used for development        |
-| netcdf4       | ^1.6.2            | To read and write NetCDF and HDF5 files               |
-| hydra-core    | ^1.3.2            | For managing application configurations               |
-| mesa-geo      | ^0.5.0            | To create spatially explicit agent-based models       |
-| xarray        | ^2023.8.0         | To work with labelled multi-dimensional arrays        |
-| fiona         | ^1.9.4.post1      | For reading and writing vector data (shapefiles, etc) |
-| fiona         | ^1.9.4.post1      | For reading and writing vector data (shapefiles, etc) |
+| python        | ">=3.9,<3.12"          | Core programming language used for development        |
+| netcdf4       | ">=1.6"           | To read and write NetCDF and HDF5 files               |
+| hydra-core    | "~1.3"           | For managing application configurations               |
+| mesa-geo      | "^0.6.0"            | To create spatially explicit agent-based models       |
+| xarray        | "~2023"        | To work with labelled multi-dimensional arrays        |
+| fiona         | ">1.8"     | For reading and writing vector data (shapefiles, etc) |
+| loguru         | "~0.7"     | For better logging |
+| rioxarray         | "~0.13"     | Operating raster data and xarray |
+| pendulum         | "~2"     | For time control |
+| geopandas         | "~0"     | For shapefile geo-data operating |
