### 2024-01-11 632975a
docs(docs): :memo: update changelog for v0.4.0
