+++ requirements.txt
@@ -44,7 +44,7 @@ ipyvue==1.10.1 ; python_version >= "3.9" and python_version < "3.12"
 ipyvuetify==1.8.10 ; python_version >= "3.9" and python_version < "3.12"
 ipywidgets==8.1.1 ; python_version >= "3.9" and python_version < "3.12"
 jedi==0.19.1 ; python_version >= "3.9" and python_version < "3.12"
-jinja2==3.1.2 ; python_version >= "3.9" and python_version < "3.12"
+jinja2==3.1.3 ; python_version >= "3.9" and python_version < "3.12"
 jsonschema-specifications==2023.12.1 ; python_version >= "3.9" and python_version < "3.12"
 jsonschema==4.20.0 ; python_version >= "3.9" and python_version < "3.12"
 jupyter-client==8.6.0 ; python_version >= "3.9" and python_version < "3.12"
@@ -54,7 +54,7 @@ kiwisolver==1.4.5 ; python_version >= "3.9" and python_version < "3.12"
 libpysal==4.8.1 ; python_version >= "3.9" and python_version < "3.12"
 loguru==0.7.2 ; python_version >= "3.9" and python_version < "3.12"
 markdown-it-py==3.0.0 ; python_version >= "3.9" and python_version < "3.12"
-markdown==3.5.1 ; python_version >= "3.9" and python_version < "3.12"
+markdown==3.5.2 ; python_version >= "3.9" and python_version < "3.12"
 markupsafe==2.1.3 ; python_version >= "3.9" and python_version < "3.12"
 matplotlib-inline==0.1.6 ; python_version >= "3.9" and python_version < "3.12"
 matplotlib==3.8.2 ; python_version >= "3.9" and python_version < "3.12"
