### 2024-01-11 da65e7c
docs(project): :memo: update project readme
