+++ docs/home/license.md
@@ -2,18 +2,8 @@
 title: license
 authors: SongshGeo
 date: 2023-01-10
-long_name: Agent-Based Social-ecological systems Modelling Framework in Python
-name: AB-SESpy
-state: open
-banner_icon: 💻
-banner: "https://songshgeo-picgo-1302043007.cos.ap-beijing.myqcloud.com/uPic/abses_github_repo.svg"
-banner_y: 0.52
-SES: social-ecological systems
-github: https://github.com/SongshGeo
-email: songshgeo@gmail.com
-website: https://cv.songshgeo.com
 ---
-# License
+## License
 
 Copyright 2023, `ABSESpy` [Shuang Song](https://cv.songshgeo.com/)
 
