+++ docs/wiki/about.md
@@ -1,20 +1,5 @@
 ---
 title: about
-authors: SongshGeo
-date: 2023-01-10
-long_name: Agent-Based Social-ecological systems Modelling Framework in Python
-name: AB-SESpy
-state: finished
-banner_icon: 💻
-banner: "https://songshgeo-picgo-1302043007.cos.ap-beijing.myqcloud.com/uPic/abses_github_repo.svg"
-banner_y: 0.52
-SES: social-ecological systems
-github: https://github.com/SongshGeo
-email: songshgeo@gmail.com
-website: https://cv.songshgeo.com
-
-emergence: ../background#emergence
-cas: ../background#complex-adaptive-systems#complex-adaptive-systems(cas)
 ---
 ## What is an Agent-based Model (ABM)
 
@@ -32,7 +17,7 @@ ABM is simulated over time: agents make decisions and act based on their attribu
 
 ## ABM is an essential method in studying SES
 
-**Several characteristics of AMB make it an essential method for SES research:**
+**Several characteristics of ABM make it an essential method for SES research:**
 
 1. It focuses on the change of an SES over time from mutual adaptations of agents and their environments.
 2. It can lead to [emergence], -system-level outcomes from micro-level interactions and macro-level feedback.
