### 2024-01-11 6816e58
ci(project): :green_heart: fix mkdocs deploy CI
