### 2024-01-11 7ec45d1
test: :white_check_mark: fix datacollector collector to str, refactoring tests and passing them
