### 2024-01-11 c7ed054
chore: :bookmark: update changelog to 0.4.2
