### 2024-01-11 f9da0ce
chore: :bookmark: update changelog to 0.4.2
