### 2024-01-14 36c5582
Updated movement notebook with more context
