### 2024-01-22 51d2bdf
Advances on second tutorials and more on logging
