+++ docs/tutorial/tutorial.md
@@ -2,7 +2,7 @@
 
 <!-- 这里，我们提供了一系列基于Jupyter Notebook的教程，展示实际利用 ABSESpy 开发可能碰到的需求及其解决策略。 -->
 
-Here, we provide a series of tutorials based on [Jupyter Notebook], demonstrating the potential needs and solutions when developing with `ABSESpy` in practice. We expect you have read through [Quick Start] and are sure that `ABSESpy` is [the right software for you]. Three levels of tutorials are available below:
+Here, we provide a series of tutorials based on [Jupyter Notebook], demonstrating the potential needs and solutions when developing with `ABSESpy` in practice. We expect you to have read through [Quick Start] guide and that you have assessed whether `ABSESpy` is [the right software for you]. Three levels of tutorials are available below:
 
 > [!INFO]In Progress
 > This document is a work in progress if you see any errors, or exclusions or have any problems, please [get in touch with us](https://github.com/absespy/ABSESpy/issues).
