### 2024-02-05 2b3bb67
docs(docs): :memo: update geodata docs, include fire tutorial
