### 2024-02-10 ABSESpy/ABSESpy#36 @wang-boyu
docs: missing DOI and citation in `refs.bib`
Hi, I'm reviewing your JOSS paper and found some missing info in your references. More specifically -

- DOI is missing for paper "Schulze, J., Müller, B., Groeneveld, J., & Grimm, V. (2017). Agent-based modelling of social-ecological systems: Achievements, challenges, and a way forward. _Journal of Artificial Societies and Social Simulation_, 20(2), 8."
- The paper cited Mesa-Geo but it's not included in the `refs.bib` file.

On a side note, it would be nice if you could format the references is a consistent way, either all in title case, or all in sentence case, but not a mix of both. For example,

**Sentence case:** Beckage, B., Moore, F. C., & Lacasse, K. (2022).  Incorporating human behaviour into Earth system modelling. ...

**Title Case:** Beckage, B., Moore, F. C., & Lacasse, K. (2022). Incorporating Human Behaviour into Earth System Modelling. ...

But this is not strictly required; just something nice to have for your own consideration. It's not blocking the paper review.
