### 2024-02-10 #6298 @wang-boyu
Raised https://github.com/ABSESpy/ABSESpy/issues/36 for missing DOI and reference.
