### 2024-02-15 ABSESpy/ABSESpy#36 @jamesdamillington
On line 99 of the [JOSS proof](https://github.com/openjournals/joss-reviews/issues/6298) there is a reference to _wang2022_ that is missing (this may be the same reference @wang-boyu refers to in their [second bullet point](https://github.com/ABSESpy/ABSESpy/issues/36#issue-2128584274)). Please add this reference.
