### 2024-02-15 ABSESpy/ABSESpy#37 @jamesdamillington
failed installation
@SongshGeo  I have just tried to install ABSESpy into an empty mamba environment using `pip install abses`

Installation seemed to proceed well until I received the error below. Could you please help me complete installation so that I can complete my [JOSS review](https://github.com/openjournals/joss-reviews/issues/6298).

```
Collecting scipy>=1.8 (from libpysal->mesa-geo<0.7.0,>=0.6.0->abses)
  Downloading scipy-1.12.0.tar.gz (56.8 MB)
  Installing build dependencies ... done
  Installing backend dependencies ... done
  Preparing metadata (pyproject.toml) ... error
  error: subprocess-exited-with-error

  × Preparing metadata (pyproject.toml) did not run successfully.
  │ exit code: 1
  ╰─> [44 lines of output]
      + meson setup /tmp/pip-install-txdr34mr/scipy_920dc993cb864aa4b1e8dad204440bb6 /tmp/pip-install-txdr34mr/scipy_920dc993cb864aa4b1e8dad204440bb6/.mesonpy-tcl4tgtp -Dbuildtype=release -Db_ndebug=if-release -Db_vscrt=md --native-file=/tmp/pip-install-txdr34mr/scipy_920dc993cb864aa4b1e8dad204440bb6/.mesonpy-tcl4tgtp/meson-python-native-file.ini
      The Meson build system
      Version: 1.3.2
      Source dir: /tmp/pip-install-txdr34mr/scipy_920dc993cb864aa4b1e8dad204440bb6
      Build dir: /tmp/pip-install-txdr34mr/scipy_920dc993cb864aa4b1e8dad204440bb6/.mesonpy-tcl4tgtp
      Build type: native build
      Project name: scipy
      Project version: 1.12.0
      C compiler for the host machine: cc (gcc 9.4.0 "cc (Ubuntu 9.4.0-1ubuntu1~20.04.2) 9.4.0")
      C linker for the host machine: cc ld.bfd 2.34
      C++ compiler for the host machine: c++ (gcc 9.4.0 "c++ (Ubuntu 9.4.0-1ubuntu1~20.04.2) 9.4.0")
      C++ linker for the host machine: c++ ld.bfd 2.34
      Cython compiler for the host machine: cython (cython 3.0.8)
      Host machine cpu family: x86_64
      Host machine cpu: x86_64
      Program python found: YES (/home/james/.mamba/bin/python3.9)
      Run-time dependency python found: YES 3.9
      Program cython found: YES (/tmp/pip-build-env-y5sgna7h/overlay/bin/cython)
      Compiler for C supports arguments -Wno-unused-but-set-variable: YES
      Compiler for C supports arguments -Wno-unused-function: YES
      Compiler for C supports arguments -Wno-conversion: YES
      Compiler for C supports arguments -Wno-misleading-indentation: YES
      Library m found: YES
      Fortran compiler for the host machine: gfortran (gcc 9.4.0 "GNU Fortran (Ubuntu 9.4.0-1ubuntu1~20.04.2) 9.4.0")
      Fortran linker for the host machine: gfortran ld.bfd 2.34
      Compiler for Fortran supports arguments -Wno-conversion: YES
      Checking if "-Wl,--version-script" : links: YES
      Program pythran found: YES (/tmp/pip-build-env-y5sgna7h/overlay/bin/pythran)
      Found pkg-config: YES (/usr/bin/pkg-config) 0.29.1
      Did not find CMake 'cmake'
      Found CMake: NO
      Run-time dependency xsimd found: NO (tried pkgconfig and cmake)
      Run-time dependency threads found: YES
      Library npymath found: YES
      Library npyrandom found: YES
      pybind11-config found: YES (/tmp/pip-build-env-y5sgna7h/overlay/bin/pybind11-config) 2.11.1
      Run-time dependency pybind11 found: YES 2.11.1
      Run-time dependency scipy-openblas found: NO (tried pkgconfig)
      Run-time dependency openblas found: NO (tried pkgconfig and cmake)
      Run-time dependency openblas found: NO (tried pkgconfig)

      ../scipy/meson.build:169:9: ERROR: Dependency "OpenBLAS" not found, tried pkgconfig

      A full log can be found at /tmp/pip-install-txdr34mr/scipy_920dc993cb864aa4b1e8dad204440bb6/.mesonpy-tcl4tgtp/meson-logs/meson-log.txt
      [end of output]

error: metadata-generation-failed

× Encountered error while generating package metadata.
╰─> See above for output.

note: This is an issue with the package mentioned above, not pip.
hint: See above for details.
```

Full terminal output, prior to that above, as follows:
```
Collecting abses
  Downloading abses-0.3.5-py3-none-any.whl.metadata (5.2 kB)
Collecting fiona>1.8 (from abses)
  Downloading fiona-1.9.5.tar.gz (409 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 409.3/409.3 kB 3.3 MB/s eta 0:00:00
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting geopandas<1,>=0 (from abses)
  Downloading geopandas-0.14.3-py3-none-any.whl.metadata (1.5 kB)
Collecting hydra-core<1.4,>=1.3 (from abses)
  Downloading hydra_core-1.3.2-py3-none-any.whl (154 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 154.5/154.5 kB 2.2 MB/s eta 0:00:00
Collecting loguru<0.8,>=0.7 (from abses)
  Downloading loguru-0.7.2-py3-none-any.whl.metadata (23 kB)
Collecting mesa-geo<0.7.0,>=0.6.0 (from abses)
  Downloading Mesa_Geo-0.6.0-py3-none-any.whl.metadata (5.6 kB)
Collecting netcdf4>=1.6 (from abses)
  Downloading netCDF4-1.6.5.tar.gz (764 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 765.0/765.0 kB 4.3 MB/s eta 0:00:00
  Installing build dependencies ... done
  Installing backend dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting numpy>=1.24 (from abses)
  Downloading numpy-1.26.4-pp39-pypy39_pp73-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (61 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 61.0/61.0 kB 3.0 MB/s eta 0:00:00
Collecting pandas<3,>=1.4 (from abses)
  Downloading pandas-2.2.0.tar.gz (4.4 MB)
  Installing build dependencies ... done
  Installing backend dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting pendulum<3,>=2 (from abses)
  Downloading pendulum-2.1.2.tar.gz (81 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 81.2/81.2 kB 9.1 MB/s eta 0:00:00
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting rioxarray<0.14,>=0.13 (from abses)
  Downloading rioxarray-0.13.4-py3-none-any.whl (53 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 53.4/53.4 kB 7.0 MB/s eta 0:00:00
Collecting xarray<2024,>=2023 (from abses)
  Downloading xarray-2023.12.0-py3-none-any.whl.metadata (11 kB)
Collecting attrs>=19.2.0 (from fiona>1.8->abses)
  Downloading attrs-23.2.0-py3-none-any.whl.metadata (9.5 kB)
Requirement already satisfied: certifi in ./.mamba/lib/python3.9/site-packages (from fiona>1.8->abses) (2023.11.17)
Collecting click~=8.0 (from fiona>1.8->abses)
  Using cached click-8.1.7-py3-none-any.whl.metadata (3.0 kB)
Collecting click-plugins>=1.0 (from fiona>1.8->abses)
  Downloading click_plugins-1.1.1-py2.py3-none-any.whl (7.5 kB)
Collecting cligj>=0.5 (from fiona>1.8->abses)
  Downloading cligj-0.7.2-py3-none-any.whl (7.1 kB)
Collecting six (from fiona>1.8->abses)
  Using cached six-1.16.0-py2.py3-none-any.whl (11 kB)
Requirement already satisfied: setuptools in ./.mamba/lib/python3.9/site-packages (from fiona>1.8->abses) (68.2.2)
Collecting importlib-metadata (from fiona>1.8->abses)
  Downloading importlib_metadata-7.0.1-py3-none-any.whl.metadata (4.9 kB)
Requirement already satisfied: packaging in ./.mamba/lib/python3.9/site-packages (from geopandas<1,>=0->abses) (23.2)
Collecting pyproj>=3.3.0 (from geopandas<1,>=0->abses)
  Downloading pyproj-3.6.1-pp39-pypy39_pp73-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (31 kB)
Collecting shapely>=1.8.0 (from geopandas<1,>=0->abses)
  Downloading shapely-2.0.2.tar.gz (279 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 279.7/279.7 kB 7.1 MB/s eta 0:00:00
  Installing build dependencies ... done
  Installing backend dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting omegaconf<2.4,>=2.2 (from hydra-core<1.4,>=1.3->abses)
  Downloading omegaconf-2.3.0-py3-none-any.whl (79 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 79.5/79.5 kB 9.9 MB/s eta 0:00:00
Collecting antlr4-python3-runtime==4.9.* (from hydra-core<1.4,>=1.3->abses)
  Downloading antlr4-python3-runtime-4.9.3.tar.gz (117 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 117.0/117.0 kB 6.2 MB/s eta 0:00:00
  Preparing metadata (setup.py) ... done
Collecting mesa~=2.1.0 (from mesa-geo<0.7.0,>=0.6.0->abses)
  Downloading Mesa-2.1.5-py3-none-any.whl.metadata (7.1 kB)
Collecting libpysal (from mesa-geo<0.7.0,>=0.6.0->abses)
  Downloading libpysal-4.8.1-py3-none-any.whl.metadata (4.6 kB)
Collecting rtree (from mesa-geo<0.7.0,>=0.6.0->abses)
  Downloading Rtree-1.2.0-py3-none-manylinux2014_x86_64.manylinux_2_17_x86_64.whl.metadata (2.1 kB)
Collecting rasterio (from mesa-geo<0.7.0,>=0.6.0->abses)
  Downloading rasterio-1.3.9.tar.gz (411 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 411.7/411.7 kB 5.6 MB/s eta 0:00:00
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting folium (from mesa-geo<0.7.0,>=0.6.0->abses)
  Downloading folium-0.15.1-py2.py3-none-any.whl.metadata (3.4 kB)
Collecting xyzservices>=2022.9.0 (from mesa-geo<0.7.0,>=0.6.0->abses)
  Downloading xyzservices-2023.10.1-py3-none-any.whl.metadata (4.0 kB)
Collecting cftime (from netcdf4>=1.6->abses)
  Downloading cftime-1.6.3.tar.gz (53 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 53.3/53.3 kB 5.6 MB/s eta 0:00:00
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting python-dateutil>=2.8.2 (from pandas<3,>=1.4->abses)
  Using cached python_dateutil-2.8.2-py2.py3-none-any.whl (247 kB)
Collecting pytz>=2020.1 (from pandas<3,>=1.4->abses)
  Downloading pytz-2024.1-py2.py3-none-any.whl.metadata (22 kB)
Collecting tzdata>=2022.7 (from pandas<3,>=1.4->abses)
  Downloading tzdata-2024.1-py2.py3-none-any.whl.metadata (1.4 kB)
Collecting pytzdata>=2020.1 (from pendulum<3,>=2->abses)
  Downloading pytzdata-2020.1-py2.py3-none-any.whl (489 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 490.0/490.0 kB 1.7 MB/s eta 0:00:00
Collecting cookiecutter (from mesa~=2.1.0->mesa-geo<0.7.0,>=0.6.0->abses)
  Downloading cookiecutter-2.5.0-py3-none-any.whl.metadata (7.3 kB)
Collecting matplotlib (from mesa~=2.1.0->mesa-geo<0.7.0,>=0.6.0->abses)
  Downloading matplotlib-3.8.3-pp39-pypy39_pp73-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (5.8 kB)
Collecting mesa-viz-tornado>=0.1.3,~=0.1.0 (from mesa~=2.1.0->mesa-geo<0.7.0,>=0.6.0->abses)
  Downloading Mesa_Viz_Tornado-0.1.3-py3-none-any.whl.metadata (1.3 kB)
Collecting networkx (from mesa~=2.1.0->mesa-geo<0.7.0,>=0.6.0->abses)
  Downloading networkx-3.2.1-py3-none-any.whl.metadata (5.2 kB)
Collecting solara (from mesa~=2.1.0->mesa-geo<0.7.0,>=0.6.0->abses)
  Downloading solara-1.26.1-py2.py3-none-any.whl.metadata (9.6 kB)
Requirement already satisfied: tqdm in ./.mamba/lib/python3.9/site-packages (from mesa~=2.1.0->mesa-geo<0.7.0,>=0.6.0->abses) (4.66.1)
Collecting PyYAML>=5.1.0 (from omegaconf<2.4,>=2.2->hydra-core<1.4,>=1.3->abses)
  Downloading PyYAML-6.0.1.tar.gz (125 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 125.2/125.2 kB 4.3 MB/s eta 0:00:00
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting affine (from rasterio->mesa-geo<0.7.0,>=0.6.0->abses)
  Downloading affine-2.4.0-py3-none-any.whl (15 kB)
Collecting snuggs>=1.4.1 (from rasterio->mesa-geo<0.7.0,>=0.6.0->abses)
  Downloading snuggs-1.4.7-py3-none-any.whl (5.4 kB)
Collecting branca>=0.6.0 (from folium->mesa-geo<0.7.0,>=0.6.0->abses)
  Downloading branca-0.7.1-py3-none-any.whl.metadata (1.5 kB)
Collecting jinja2>=2.9 (from folium->mesa-geo<0.7.0,>=0.6.0->abses)
  Downloading Jinja2-3.1.3-py3-none-any.whl.metadata (3.3 kB)
Requirement already satisfied: requests in ./.mamba/lib/python3.9/site-packages (from folium->mesa-geo<0.7.0,>=0.6.0->abses) (2.31.0)
Collecting zipp>=0.5 (from importlib-metadata->fiona>1.8->abses)
  Using cached zipp-3.17.0-py3-none-any.whl.metadata (3.7 kB)
Requirement already satisfied: platformdirs>=2.0.2 in ./.mamba/lib/python3.9/site-packages (from libpysal->mesa-geo<0.7.0,>=0.6.0->abses) (4.1.0)
Collecting beautifulsoup4>=4.10 (from libpysal->mesa-geo<0.7.0,>=0.6.0->abses)
  Downloading beautifulsoup4-4.12.3-py3-none-any.whl.metadata (3.8 kB)
Collecting scipy>=1.8 (from libpysal->mesa-geo<0.7.0,>=0.6.0->abses)
  Downloading scipy-1.12.0.tar.gz (56.8 MB)
```
