### 2024-02-15 ABSESpy/ABSESpy#37 @SongshGeo
@jamesdamillington  I sincerely apologize for the installation issues you've encountered with ABSESpy. I will promptly investigate the matter to understand and resolve the problem. Please note that we are currently in the midst of the Chinese New Year holiday, which may affect my working hours. Despite this, I assure you that I will provide you with feedback within the next 5 days. Thank you for your understanding and patience.
