### 2024-02-18 ABSESpy/ABSESpy#37 @SongshGeo
@jamesdamillington

Hi, thanks for waiting. However, I can't reproduce this issue on my MacOS with an empty Virenv, unfortunately. I guess this is a platform-dependent issue. Please tell me more about your environment; I will try to reproduce the issue under a similar environment. What's more, it seems not an issue of `ABSESpy` but -as shown- by `scipy`.

I found [this issue](https://github.com/scipy/scipy/issues/16308) from their repo. Please check if the recommended solutions can solve this issue. :)
