+++ docs/refs.bib
@@ -1,5 +1,17 @@
+@book{arthur2018,
+  title = {The Economy as an Evolving Complex System II},
+  author = {Arthur, W. Brian},
+  year = {2018},
+  month = may,
+  edition = {1},
+  publisher = {CRC Press},
+  doi = {10.1201/9780429496639},
+  isbn = {978-0-429-49663-9},
+  language = {english}
+}
+
 @article{beck2011,
-  title = {On Governance for Re-Engineering City Infrastructure},
+  title = {On governance for re-engineering city infrastructure},
   author = {Beck, M. Bruce and Thompson, Michael and Ney, Steven and Gyawali, Dipak and Jeffrey, Paul},
   year = {2011},
   month = jun,
@@ -7,279 +19,25 @@
   volume = {164},
   number = {2},
   pages = {129--142},
-  publisher = {{ICE Publishing}},
+  publisher = {ICE Publishing},
   issn = {1478-4629},
-  doi = {10.1680/ensu.2011.164.2.129},
-  abstract = {Given the challenge of re-engineering the infrastructure of a city so that it may become a force for good in the environment, what structure of governance might promote/stifle any associated technological and technical innovations? How might the `quality' of the sought-for governance be gauged and improved? Cultural theory is introduced to address these questions and briefly elaborated. In particular, it is shown how cultural theory can be used to refurbish Dahl's theory of pluralist democracy and provide a simple (and visual) metric of deliberative quality in governance. Matters of re-engineering the water-nutrient infrastructures of the city of Atlanta within the Chattahoochee watershed in the southeastern USA provide the primary setting for eventually applying the resulting conceptual framework. Drawing together a somewhat eclectic mix of past experience and prospective approaches \textendash{} ranging from practical resolutions of water conflicts in Nepal, lessons learned from previous community engagement in the Atlanta\textendash Chattahoochee system, and two stagings of the sustainability agora of the International Water Association \textendash{} a cultural theory-inspired programme of work is developed to explore and assemble those elements of governance that should, in practice, enable the re-engineering needed for a city to become a force for good in its environment.},
-  keywords = {\#SHModel_YRB_2022,⛓️关联,plural rationality theory,to-read},
-  file = {/Users/songshgeo/Zotero/storage/UCWLIXUR/Beck et al_2011_On governance for re-engineering city infrastructure.pdf}
+  doi = {10.1680/ensu.2011.164.2.129}
 }
 
 @incollection{beck2018,
-  title = {Applying {{Plural Rationality}} to {{Some Wicked Problems}}},
-  shorttitle = {Societal {{Drivers}} of {{Food}} and {{Water Systems}} 2},
-  booktitle = {Societal {{Drivers}} of {{Food}} and {{Water Systems}}},
+  title = {Applying Plural Rationality to Some Wicked Problems},
+  shorttitle = {Societal Drivers of Food and Water Systems 2},
+  booktitle = {Societal Drivers of Food and Water Systems},
   author = {Beck, M. B. and Gyawali, D. and Thompson, M.},
   editor = {Allan, T. and Bromwich, B. and Colman, A. and Keulertz, M.},
   year = {2018},
-  publisher = {{Oxford University Press}},
-  address = {{Oxford, UK}},
[…]
