### 2024-02-18 ABSESpy/ABSESpy#36 @SongshGeo
Thanks for the explanation by @jamesdamillington; yep, the reference to `mesa-geo` was not rendered because it lacked a `bib` item. It would be arrogant if I totally ignored `mesa-geo` but invited @wang-boyu  to review this repo, LOL.

The behavior of the case-sensitive problem was because when exporting the `.bib` file, a 'keep case' option was selected.

Now everything is fixed in PR #38.
