+++ docs/tutorial/tutorial.md
@@ -37,6 +37,9 @@ You're looking to create and operate a basic model by `ABSESpy` framework. Wonde
 [:material-book: __B08__](beginner/fire_tutorial.ipynb) A classic agent-based model.
 { .card }
 
+[:material-book: __B09__](beginner/predation_tutorial.ipynb) Wolf-Sheep predation model.
+{ .card }
+
 > :material-book: __Huh__ ... More tutorials in this level are coming soon.
 
 </div>
