+++ Fire spread model.md
@@ -1,5 +1,0 @@
-I would need to update a 12 by 12 array based whether there is an agent on that particular path and on whether there agent on that path is burning or another state
-
-
----
-I need to get some network data analysis done this weekend so not a ton of sleep.  Network data analysis done with igraph MAYBE and MAYBE locally just that I bring calculation later into datalore such that I can present results. Now, think
