### 2024-02-29 b553e55
feat(src): :sparkles: Random choose now can select from an empty list
use when_empty arg to control the behavior
