### 2024-02-29 25bdf5b
refactor(docs): :memo: predation model is now more concise, fix blank lines
