### 2024-02-29 5e37ed2
fix(src): :bug: alter nature now behaves correctly
