### 2024-02-29 ABSESpy/ABSESpy#37 @jamesdamillington
I think I have now been able to install `abses` via pip. Before running `pip install abses` I needed to [install](https://www.linuxcapable.com/how-to-install-cmake-on-ubuntu-linux/) a recent version of `cmake` and to install `libopenblas-dev` `libopenblas64-dev` (as these were also [not being found](https://stackoverflow.com/a/73782008)).

However, now when trying to `import abses` in python I get the error:
```
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/james/.mamba/lib/pypy3.9/site-packages/abses/__init__.py", line 33, in <module>
    from .actor import Actor, perception
  File "/home/james/.mamba/lib/pypy3.9/site-packages/abses/actor.py", line 10, in <module>
    from typing import (
ImportError: cannot import name 'TypeAlias' from 'typing' (/home/james/.mamba/lib/pypy3.9/typing.py)
```
Other installed python packages (like `numpy` and `scipy`) load without problem.

Is this problem because I haven't
> properly configured the environment to import the module into your workspace

as indicated on [the Getting Started page](https://absespy.github.io/ABSESpy/home/get_started/)? What do I need to do to 'properly configure' my environment?
