### 2024-02-29 ABSESpy/ABSESpy#37 @SongshGeo
@jamesdamillington  I'm glad the `absespy` was installed successfully. Regarding the newest import error, it was caused by a Python version-dependent problem. An update to Python 3.10 is expected to fix this problem.

However, `absespy` was designed to be compatible with Python 3.9. Obviously it failed to do so now, as codes kept growing... I'm sorry about this incompatible issue and we will fix this as soon as possible.
