### 2024-03-04 1fa9c2c
refactor(actors): :recycle: actors' movement by new proxy class
