+++ abses/actor.py
@@ -6,7 +6,7 @@
 # Website: https://cv.songshgeo.com/
 from __future__ import annotations
 
-from functools import wraps
+from functools import cached_property, wraps
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -14,18 +14,16 @@ from typing import (
     Dict,
     Iterable,
     Optional,
-    Tuple,
+    Self,
     TypeAlias,
     Union,
 )
 
 import mesa_geo as mg
-from mesa.space import Coordinate
-from shapely import Point
 
 from abses.decision import DecisionFactory
-from abses.errors import ABSESpyError
 from abses.links import LinkNode
+from abses.move import _Movements
 from abses.objects import _BaseObj
 from abses.sequences import ActorsList
 from abses.tools.func import make_list
@@ -34,6 +32,7 @@ from abses.tools.func import make_list
 
 
 if TYPE_CHECKING:
+    from abses.human import LinkContainer
     from abses.main import MainModel
     from abses.nature import PatchCell
 
@@ -109,14 +108,6 @@ class Actor(mg.GeoAgent, _BaseObj, LinkNode):
             Whether the actor is standing on a cell.
         here:
             A list of actors that are on the same cell as the actor.
-
-    Methods:
-        put_on(self, cell: PatchCell | None = None) -> None
-            Places the actor on a cell.
-        put_on_layer(self, layer: mg.RasterLayer, pos: Tuple[int, int])
-            Specifies a new cell for the actor to be located on.
-        selecting(self, selection: Union[str, Dict[str, Any]]) -> bool
-            Selects the actor according to specified criteria.
     """
 
     # when checking the rules
@@ -139,10 +130,9 @@ class Actor(mg.GeoAgent, _BaseObj, LinkNode):
             self, unique_id, model=model, geometry=geometry, crs=crs
         )
         LinkNode.__init__(self)
-        self._rules: Dict[str, Dict[str, Any]] = {}
         self._cell: PatchCell = None
-        self.container = model.human
-        self._decisions = self._setup_decisions()
+        self._container: LinkContainer = model.human
+        self._decisions: DecisionFactory = self._setup_decisions()
 
     def _setup_decisions(self) -> DecisionFactory:
         """Decisions that this actor makes."""
@@ -163,26 +153,6 @@ class Actor(mg.GeoAgent, _BaseObj, LinkNode):
         return None if self._cell is None else self._cell.layer
 
     @property
-    def indices(self) -> Coordinate:
-        """Coordinates in the form of (row, col) are used for indexing a matrix, with the origin at the top left corner and increasing downwards."""
-        return self._cell.indices
-
-    @property
-    def pos(self) -> Coordinate:
-        """Coordinates in the form of (x, y) indexed from the bottom left corner."""
-        return self._cell.pos
-
-    @pos.setter
-    def pos(self, pos) -> None:
-        if pos is not None:
-            raise AttributeError(f"Set pos by {self.put_on_layer.__name__}")
-
-    @property
-    def population(self) -> ActorsList:
-        """List of agents of the same breed"""
-        return self.model.agents[self.breed]
-
-    @property
     def on_earth(self) -> bool:
         """Whether agent stands on a cell"""
         return bool(self._cell)
@@ -192,57 +162,40 @@ class Actor(mg.GeoAgent, _BaseObj, LinkNode):
         """Other agents on the same cell as the agent."""
         return self._cell.agents
 
-    def put_on(self, cell: PatchCell | None = None) -> None:
-        """Place agent on a cell. If the agent is already located at a cell, it should be located to a cell with the same layer.
+    @property
+    def at(self) -> PatchCell | None:
+        """Get the cell where the agent is located."""
+        return self._cell if self._cell is not None else None
[…]
