### 2024-03-04 d6a387c
refactor(src): :recycle: AgentsContainer private and not singleton anymore
