### 2024-03-04 6055e78
refactor(src): :recycle: store agents by container in PatchCell
