### 2024-03-05 9d01755
fix(random): :bug: fixing `replace` arg doesn't work problem
