+++ tests/fixtures.py
@@ -31,22 +31,6 @@ class Nature(BaseNature):
         return data[list(cfg.crops_id)]
 
 
-class Farmer(Actor):
-    """测试用，另一个类别的主体"""
-
-    def __init__(self, model, observer: bool = True) -> None:
-        super().__init__(model, observer)
-        self.metric = 0.1
-
-
-class Admin(Actor):
-    """测试用，另一个类别的主体"""
-
-
-class City(Actor):
-    """测试用，每个城市的主体"""
-
-
 @pytest.fixture(name="water_quota_model")
 def setup_water_quota_model() -> MainModel:
     """创造可供测试的黄河灌溉用水例子"""
