+++ tests/test_container.py
@@ -5,31 +5,143 @@
 # GitHub   : https://github.com/SongshGeo
 # Website: https://cv.songshgeo.com/
 
+from typing import Dict, List
 
-from abses import MainModel
+import pytest
 
-from .fixtures import Actor, Admin, Farmer
+from abses import Actor, MainModel
+from abses.container import _AgentsContainer
 
 
-def test_container_attributes():
-    """测试容器的属性"""
-    model = MainModel()
-    container = model.agents
-    assert model.agents is container
-    assert len(container) == 0
-    assert repr(container) == "<AgentsContainer: >"
-    assert container.model is model
-    a_farmer = container.create(Farmer, singleton=True)
-    admins_5 = container.create(Admin, 5)
-    assert isinstance(a_farmer, Actor)
-    assert len(container) == 6
-    assert repr(container) == "<AgentsContainer: (1)Farmer; (5)Admin>"
-    assert container.Admin == admins_5
+class TestMainContainer:
+    """
+    测试用于整个模型的主体容器。
+    """
 
-    # 增删
-    another_farmer = Farmer(model)
-    assert "Farmer" in container.keys()
-    container.add(another_farmer)
-    container.remove(admins_5[0])
-    admins_5[1:3].trigger("die")
-    assert repr(container) == "<AgentsContainer: (2)Farmer; (2)Admin>"
+    @pytest.fixture(name="model", scope="function")
+    def mock_model(self):
+        """创建一个模型"""
+        return MainModel()
+
+    @pytest.fixture(name="module", scope="function")
+    def mock_module(self, model):
+        """创建一个（2*2）的斑块模块"""
+        return model.nature.create_module(how="from_resolution", shape=(2, 2))
+
+    def test_register(self, model, module):
+        """测试注册，注册的主体类型应该在模型的所有 Container 中都自动被注册。"""
+        # arrange
+        container = model.agents
+
+        # action
+        container.register([Actor])
+
+        # assert
+        assert all("Actor" in cell.agents.keys() for cell in module)
+        assert repr(container) == "<ModelAgents: (0)Actor>"
+
+    def test_create(self, model):
+        """测试创建主体"""
+        # arrange
+        container: _AgentsContainer = model.agents
+
+        # action
+        actor = container.create(Actor, singleton=True)
+
+        # assert
+        assert actor in container.Actor
+        assert repr(container) == "<ModelAgents: (1)Actor>"
+
+    @pytest.mark.parametrize(
+        "init_agent, agent_num, expected_repr",
+        [
+            (["Admin"], [3], "<ModelAgents: (3)Admin>"),
+            (
+                ["Admin", "Farmer"],
+                [3, 2],
+                "<ModelAgents: (3)Admin; (2)Farmer>",
+            ),
+            ([], [], "<ModelAgents: >"),
+            (
+                ["Admin", "Farmer"],
+                [0, 0],
+                "<ModelAgents: (0)Admin; (0)Farmer>",
+            ),
+        ],
+        ids=[
+            "One breed",
+            "Two breeds",
+            "Empty container",
+            "Empty container but registered two",
+        ],
+    )
+    def test_report(self, model, breeds, init_agent, agent_num, expected_repr):
+        """测试汇报模型的样子"""
+        # arrange
+        container: _AgentsContainer = model.agents
+        # action
+        for a, n in zip(init_agent, agent_num):
+            breed_cls = breeds.get(a)
+            container.create(breed_cls, n)
+        # assert
+        assert repr(container) == expected_repr
+
+    @pytest.mark.parametrize(
+        "init_breeds",
+        [
+            [],
+            ["Admin"],
+            ["Admin", "Farmer"],
+        ],
+        ids=[
[…]
