+++ tests/test_human.py
@@ -8,9 +8,6 @@
 import numpy as np
 
 from abses import MainModel
-from abses.nature import PatchModule
-
-from .fixtures import Admin, Farmer
 
 selection = {
     "breed": "Farmer",
@@ -18,21 +15,23 @@ selection = {
 }
 
 
-def test_human_attributes():
+def test_human_attributes(farmer_cls, admin_cls):
+    """测试人类模块的属性"""
     model = MainModel()
     human = model.human
     assert human.agents is model.agents
     assert len(human.actors()) == 0
-    human.agents.create(Farmer, 5)
-    model.agents.create(Admin, 5)
+    human.agents.create(farmer_cls, 5)
+    model.agents.create(admin_cls, 5)
     assert len(human.agents) == 10
 
 
-def test_human_define():
+def test_human_define(farmer_cls, admin_cls):
+    """测试人口的定义"""
     model = MainModel()
     human = model.human
-    farmers = human.agents.create(Farmer, 5)
-    admins = model.agents.create(Admin, 5)
+    farmers = human.agents.create(farmer_cls, 5)
+    admins = model.agents.create(admin_cls, 5)
 
     farmers.update("test", np.arange(5))
     admins.update("test", np.arange(5))
