### 2024-03-05 4587293
refactor(src): :recycle: separate _CellAgentsContainer and AgentsContainer
