### 2024-03-05 cdf4224
test(src): :white_check_mark: passing tests of actor's movement
