+++ tests/test_move.py
@@ -1,0 +1,144 @@
+#!/usr/bin/env python 3.11.0
+# -*-coding:utf-8 -*-
+# @Author  : Shuang (Twist) Song
+# @Contact   : SongshGeo@gmail.com
+# GitHub   : https://github.com/SongshGeo
+# Website: https://cv.songshgeo.com/
+
+from __future__ import annotations
+
+from typing import Tuple
+
+import pytest
+
+from abses.actor import Actor
+from abses.cells import PatchCell
+from abses.errors import ABSESpyError
+from abses.main import MainModel
+
+
+class TestMovements:
+    """
+    Test the movements of agents.
+    """
+
+    def test_move_to_cell(self, cell_0_0: PatchCell, cell_0_1: PatchCell):
+        """Test the movement of agents."""
+        # arrange
+        actor = cell_0_1.agents.create(Actor, singleton=True)
+        # action
+        actor.move.to(cell_0_0)
+        # assert
+        assert actor.at is cell_0_0
+        assert actor not in cell_0_1.agents
+        assert cell_0_1.agents.is_empty
+
+    def test_move_to_coordinate(
+        self, model: MainModel, cell_0_1: PatchCell, cell_0_0: PatchCell
+    ):
+        """Test the movement of agents."""
+        # arrange
+        actor = model.agents.create(Actor, singleton=True)
+        # action
+        actor.move.to(cell_0_1.pos, cell_0_1.layer)
+        # assert
+        assert actor.at is cell_0_1
+        actor.move.to((0, 0))
+        assert actor.at is cell_0_0
+        assert actor not in cell_0_1.agents
+
+    def test_move_cross_layer(self, model: MainModel, cell_0_0: PatchCell):
+        """Test raises error when move wrongly."""
+        # arrange
+        new_module = model.nature.create_module(
+            how="from_resolution", shape=(1, 2)
+        )
+        actor = cell_0_0.agents.create(Actor, singleton=True)
+        another_layer_cell = new_module.cells[0][0]
+        # action
+        with pytest.raises(ABSESpyError):
+            actor.move.to(another_layer_cell)
+        assert actor.at is cell_0_0
+        cell_0_0.agents.remove(actor)
+        actor.move.to(another_layer_cell)
+        assert actor.at is another_layer_cell
+
+    def test_move_off(self, cell_0_0: PatchCell):
+        """Test the movement of agents."""
+        # arrange
+        actor = cell_0_0.agents.create(Actor, singleton=True)
+        # action
+        actor.move.off()
+        # assert
+        assert actor.at is None
+        assert cell_0_0.agents.is_empty
+
+    @pytest.mark.parametrize(
+        "location, direction, distance, expected_location",
+        [
+            ((0, 0), "up", 1, (1, 0)),
+            ((0, 0), "down", 1, (1, 0)),
+            ((0, 0), "left", 1, (0, 1)),
+            ((0, 0), "right", 1, (0, 1)),
+            ((0, 0), "left", 2, (0, 0)),
+            ((0, 0), "up", 2, (0, 0)),
+        ],
+    )
+    def test_move_by(
+        self,
+        cells,
+        location: Tuple[int, int],
+        direction: str,
+        distance: int,
+        expected_location: tuple[int, int],
+    ):
+        """Test the movement of agents."""
+        # arrange
+        cell: PatchCell = cells[*location]
+        actor = cell.agents.create(Actor, singleton=True)
+        # action
+        actor.move.by(direction=direction, distance=distance)
+        # assert
+        new_cell: PatchCell = cells[*expected_location]
+        assert (actor in cell.agents) is (new_cell is cell)
[…]
