### 2024-03-05 8a8fd9d
fix(src): :bug: fixing actor.die
