+++ abses/actor.py
@@ -248,7 +248,6 @@ class Actor(mg.GeoAgent, _BaseObj, LinkNode):
             self._container.get_graph(link).remove_node(self)
         if self.on_earth:
             self.move.off()
-            self._cell.agents.remove(self)
             del self
 
     def alter_nature(self, attr: str, value: Any) -> None:
