### 2024-03-06 5562d09
refactor(src): :recycle: refactoring storage of linking data, converting to networkx when necessary
