### 2024-03-06 c3171a9
refactor(src): :recycle: use 'get', 'set' method to control the actor's behaviors
