+++ abses/actor.py
@@ -21,9 +21,9 @@ from typing import (
 
 import mesa_geo as mg
 
-from abses.decision import DecisionFactory
+from abses.decision import _DecisionFactory
 from abses.errors import ABSESpyError
-from abses.links import LinkNode
+from abses.links import _LinkNode
 from abses.move import _Movements
 from abses.objects import _BaseObj
 from abses.sequences import ActorsList
@@ -40,27 +40,7 @@ if TYPE_CHECKING:
 
 Selection: TypeAlias = Union[str, Iterable[bool]]
 Trigger: TypeAlias = Union[Callable, str]
-
-
-def parsing_string_selection(selection: str) -> Dict[str, Any]:
-    """Parses a string selection expression and returns a dictionary of key-value pairs.
-
-    Parameters:
-        selection:
-            String specifying which breeds to select.
-
-    Returns:
-        selection_dict:
-            Parsed output as Dictionary
-    """
-    selection_dict = {}
-    if "==" not in selection:
-        return {"breed": selection}
-    expressions = selection.split(",")
-    for exp in expressions:
-        left, right = tuple(exp.split("=="))
-        selection_dict[left.strip(" ")] = right.strip(" ")
-    return selection_dict
+Targets: TypeAlias = Union["ActorsList", "Self", "PatchCell"]
 
 
 def perception_result(name, result, nodata: Any = 0.0) -> Any:
@@ -90,7 +70,7 @@ def perception(
     return decorator(decorated_func) if decorated_func else decorator
 
 
-class Actor(mg.GeoAgent, _BaseObj, LinkNode):
+class Actor(mg.GeoAgent, _BaseObj, _LinkNode):
     """
     An actor in a social-ecological system (or "Agent" in an agent-based model.)
 
@@ -130,18 +110,18 @@ class Actor(mg.GeoAgent, _BaseObj, LinkNode):
         mg.GeoAgent.__init__(
             self, unique_id, model=model, geometry=geometry, crs=crs
         )
-        LinkNode.__init__(self)
+        _LinkNode.__init__(self)
         self._cell: PatchCell = None
         self._container: LinkContainer = model.human
-        self._decisions: DecisionFactory = self._setup_decisions()
+        self._decisions: _DecisionFactory = self._setup_decisions()
 
-    def _setup_decisions(self) -> DecisionFactory:
+    def _setup_decisions(self) -> _DecisionFactory:
         """Decisions that this actor makes."""
         decisions = make_list(getattr(self, "__decisions__", None))
-        return DecisionFactory(self, decisions)
+        return _DecisionFactory(self, decisions)
 
     @property
-    def decisions(self) -> DecisionFactory:
+    def decisions(self) -> _DecisionFactory:
         """Decisions that this agent makes."""
         return self._decisions
 
@@ -193,11 +173,22 @@ class Actor(mg.GeoAgent, _BaseObj, LinkNode):
         """Manipulating agent's location."""
         return _Movements(self)
 
+    def _get_correct_target(self, target: Targets, attr: str) -> Targets:
+        """Which targets should be used when getting or setting."""
+        if target is not None:
+            return target
+        if hasattr(self, attr):
+            return self
+        if self.on_earth:
+            return self._cell
+        raise AttributeError(f"{attr} not found in {self}.")
+
     def get(
         self,
         attr: str,
         aggfunc: Optional[Callable] = None,
         target: Optional[Self | PatchCell] = None,
+        **kwargs,
     ) -> Any:
         """Gets attribute value from target.
         attr:
@@ -209,62 +200,24 @@ class Actor(mg.GeoAgent, _BaseObj, LinkNode):
             1. If the target is an agent, get the attribute from the agent.
[…]
