### 2024-03-06 #6298 @jamesdamillington
@crvernon  I've been struggling to get things running on my linux machine. Initially it was an issue at my end but now @SongshGeo thinks it might be an issue with python versioning and is investigating, so I'm waiting for the result of that. Have I got that right @SongshGeo ?
