+++ abses/links.py
@@ -7,63 +7,113 @@
 
 from __future__ import annotations
 
-from typing import Dict, Iterable, Iterator, List, Optional, Self, Set
+from typing import (
+    TYPE_CHECKING,
+    Dict,
+    Iterable,
+    List,
+    Optional,
+    Set,
+    TypeAlias,
+)
+
+from abses.sequences import ActorsList
+
+# from abses.objects import _LinkNode
+
+if TYPE_CHECKING:
+    from abses import MainModel
+    from abses.actor import Actor
+    from abses.cells import PatchCell
+
+Node: TypeAlias = "Actor | PatchCell"
 
 
-class _LinkNode:
-    """节点类"""
+class _LinkContainer:
+    """Container for links."""
 
     def __init__(self) -> None:
-        self._linking_me: Dict[str, Set[Self]] = {}
-        self._linking_to: Dict[str, Set[Self]] = {}
-
-    @classmethod
-    @property
-    def breed(cls) -> str:
-        """种类"""
-        return cls.__name__
+        self._back_links: Dict[str, Dict[int, Set]] = {}
+        self._links: Dict[str, Dict[int, Set]] = {}
+        self._cached_networks: Dict[str, object] = {}
 
     @property
-    def links(self) -> Set[str]:
-        """链接"""
-        return tuple(self._linking_to.keys())
+    def links(self) -> tuple[str]:
+        """Get the links of a certain type."""
+        return tuple(self._links.keys())
 
-    def is_linked_by(self, node: _LinkNode, link_name: Optional[str]) -> bool:
-        """是否被连接"""
-        return node in self._linking_me.get(link_name, set())
+    @links.setter
+    def links(self, link_name: str) -> None:
+        """Set the links."""
+        self._links[link_name] = {}
+        self._back_links[link_name] = {}
 
-    def is_linking_to(self, node: _LinkNode, link_name: Optional[str]) -> bool:
-        """是否连接"""
-        return node in self._linking_to.get(link_name, set())
+    def get_graph(self, link_name):
+        """Get the graph."""
+        try:
+            import networkx as nx
+        except ImportError as exc:
+            raise ImportError(
+                "You need to install networkx to use this function."
+            ) from exc
+        graph = nx.from_dict_of_lists(self._links[link_name])
+        self._cached_networks[link_name] = graph
+        return graph
 
-    def link_to(
+    def _register_link(
+        self, link_name: str, source: _LinkNode, target: _LinkNode
+    ) -> None:
+        """Register a link."""
+        if link_name not in self._links:
+            self.links = link_name
+        if source not in self._links[link_name]:
+            self._links[link_name][source] = set()
+        if target not in self._back_links[link_name]:
+            self._back_links[link_name][target] = set()
+
+    def has_link(
+        self, link_name: str, node1: _LinkNode, node2: _LinkNode
+    ) -> tuple[bool]:
+        """If link exists."""
+        data = self._links[link_name]
+        node1_to_node2 = (
+            False if node1 not in data else node2 in data.get(node1, [])
+        )
+        node2_to_node1 = (
+            False if node2 not in data else node1 in data.get(node2, [])
+        )
+        return node1_to_node2, node2_to_node1
+
+    def add_a_link(
         self,
-        node: _LinkNode,
-        link_name: Optional[str],
+        link_name: str,
+        source: Node,
+        target: Node,
         mutual: bool = False,
     ) -> None:
-        """将行动者与其它行动者或地块建立连接"""
-        if link_name not in self._linking_to:
-            self._linking_to[link_name] = set()
-        self._linking_to[link_name].add(node)
-        # 如果对方那没有记录被自己连接，那么让它记录
[…]
