+++ abses/container.py
@@ -5,13 +5,27 @@
 # GitHub   : https://github.com/SongshGeo
 # Website: https://cv.songshgeo.com/
 
+"""
+行动者容器，集中保存行动者。
+
+是私有类，按照，应该存在于
+"""
+
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Iterable, Optional, Type, Union
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    Iterable,
+    Optional,
+    Type,
+    TypeAlias,
+    Union,
+)
 
 from loguru import logger
 
-from abses.actor import Actor
+from abses.actor import Actor, Breeds
 from abses.errors import ABSESpyError
 from abses.sequences import ActorsList, Selection
 from abses.tools.func import make_list
@@ -20,16 +34,12 @@ if TYPE_CHECKING:
     from abses.cells import PatchCell
     from abses.main import MainModel
 
-# logger = logging.getLogger("__name__")
+ActorTypes: TypeAlias = Type[Actor] | Iterable[Type[Actor]]
+Actors: TypeAlias = Union[Actor, ActorsList, Iterable[Actor]]
 
 
 class _AgentsContainer(dict):
-    """Singleton AgentsContainer for each model.
-
-    This class is a dictionary-like container for managing agents in a simulation model. It is designed to be a singleton,
-    meaning that there is only one instance of this class per model. It provides methods for creating, adding, removing,
-    and selecting agents, as well as triggering events.
-    """
+    """AgentsContainer for the main model."""
 
     def __init__(
         self,
@@ -51,13 +61,6 @@ class _AgentsContainer(dict):
         # rep = self.to_list().__repr__()[13:-1]
         strings = [f"({len(v)}){k}" for k, v in self.items()]
         return f"<{str(self)}: {'; '.join(strings)}>"
-
-    def __getattr__(self, name: str) -> Any | Actor:
-        return (
-            self.get(name)
-            if name in self.model.breeds
-            else getattr(self, name)
-        )
 
     def __contains__(self, name) -> bool:
         return name in self.get()
@@ -91,11 +94,11 @@ class _AgentsContainer(dict):
             return
         now = len(self.get())
         if now + when_adding > self._max_length:
-            raise ABSESpyError(
-                f"{self} is full (maximum {self._max_length}: Now has {now}), trying to add {when_adding} more."
-            )
+            e1 = f"{self} is full (maximum {self._max_length}: "
+            e2 = f"Now has {now}), trying to add {when_adding} more."
+            raise ABSESpyError(e1 + e2)
 
-    def register(self, actor_cls: Type[Actor] | Iterable[Type[Actor]]) -> None:
+    def register(self, actor_cls: ActorTypes) -> None:
         """Registers a new breed of actors."""
         for a_cls in make_list(actor_cls):
             breed = a_cls.breed
@@ -148,14 +151,13 @@ class _AgentsContainer(dict):
             return agents[0] if num == 1 else agents
         return agents
 
-    def get(
-        self, breeds: Optional[Union[str, Iterable[str]]] = None
-    ) -> ActorsList[Actor]:
+    def get(self, breeds: Breeds = None) -> ActorsList[Actor]:
         """Get all entities of specified breeds to a list.
 
         Parameters:
             breeds:
-                The breed(s) of entities to convert to a list (Optional[Union[str, Iterable[str]]]). If None, all breeds are used.
+                The breed(s) of entities to convert to a list.
+                If None, all breeds are used.
 
         Returns:
             ActorsList:
@@ -181,20 +183,12 @@ class _AgentsContainer(dict):
                 Keyword arguments to be passed to the `trigger` method of each agent.
 
         Returns:
-            None
[…]
