+++ abses/tools/regex.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python 3.11.0
+# -*-coding:utf-8 -*-
+# @Author  : Shuang (Twist) Song
+# @Contact   : SongshGeo@gmail.com
+# GitHub   : https://github.com/SongshGeo
+# Website: https://cv.songshgeo.com/
+
+"""
+这个模块里存储一些常用的正则表达式。
+"""
+
+# 模块名称应该符合蛇形命名法，且不能以下划线开头
+# https://www.cnblogs.com/mrlonely2018/p/15650605.html
+MODULE_NAME = r"[a-z][a-z0-9_]*"
