+++ tests/test_components.py
@@ -12,69 +12,159 @@ import pytest
 from omegaconf import DictConfig
 
 from abses.components import _Component
+from abses.main import MainModel
 
 
-class MainModel:
-    """测试模型"""
+class ArgComponent(_Component):
+    """用于测试的，包含强制参数的组件"""
 
-    def __init__(self, settings: DictConfig) -> None:
-        self.settings = settings
-
-
-def test_component_initialization():
-    """
-    测试组件的初始化
-    """
-    model = MainModel(settings={})
-    component = _Component(model, "test_component")
-    assert isinstance(component.args, DictConfig)
-    assert list(component.args.keys()) == component.__args__
+    __args__ = ["param1", "param2"]
 
 
-def test_component_params():
-    """
-    Test the initialization of a component with specific parameters.
+class TestComponent:
+    """测试模型的基本组件"""
 
-    This function creates a `MainModel` object with the given settings and then
-    creates a `Component` object using the model and the specified component name.
-    It asserts that the `params` attribute of the component is equal to the
-    expected parameters.
-    """
-    model = MainModel(
-        settings={"test_component": {"param1": "value1", "param2": "value2"}}
+    def _wrap_settings(self, settings: dict, prefix="test"):
+        """将输入的配置变成符合格式的形式"""
+        return {prefix: settings}
+
+    @pytest.mark.parametrize(
+        "module_name",
+        [
+            ("test_component"),
+            ("test_component_"),
+        ],
+        ids=[
+            "Happy path - valid name",
+            "Happy path - end with a underscore",
+        ],
     )
-    component = _Component(model, "test_component")
-    assert component.params == {"param1": "value1", "param2": "value2"}
+    def test_init_happy_path(self, model, module_name):
+        """
+        测试组件的初始化
+        """
+        # arrange / action
+        component = _Component(model, module_name)
+        # assert
+        assert isinstance(component.args, DictConfig)
+        assert list(component.args.keys()) == component.__args__
 
-
-def test_component_args_property():
-    model = MainModel(
-        settings={"test_component": {"param1": "value1", "param2": "value2"}}
+    @pytest.mark.parametrize(
+        "module_name, expected",
+        [
+            ("_component", ValueError),
+            ("TestCapitalize", ValueError),
+            (3, TypeError),
+            (("test", 111), TypeError),
+        ],
+        ids=[
+            "Invalid name - not start with a underscore",
+            "Invalid name - Capitalize",
+            "Invalid type - not string",
+            "Invalid type - not string",
+        ],
     )
-    component = _Component(model, "test_component")
-    component.add_args(["param1"])
-    assert component.args == DictConfig({"param1": "value1"})
+    def test_bad_name_init(self, model, module_name, expected):
+        """
+        测试组件的初始化
+        """
+        # arrange / action
+        with pytest.raises(expected):
+            _Component(model, module_name)
 
-
-def test_component_args_setter_string():
-    model = MainModel(settings={"test_component": {"param3": "value3"}})
-    component = _Component(model, "test_component")
-    component.add_args("param3")
-    assert "param3" in component.args
-
-
-def test_component_args_setter_iterable():
-    model = MainModel(
-        settings={"test_component": {"param4": "value4", "param5": "value5"}}
+    @pytest.mark.parametrize(
+        "settings, expected",
+        [
+            (1, 1),
+            ({"testing": 3}, {"testing": 3}),
+        ],
+        ids=[
+            "Happy path - case 1",
+            "Happy path - case 2",
+        ],
     )
-    component = _Component(model, "test_component")
-    component.add_args(["param4", "param5"])
-    assert "param4" in component.args
[…]
