### 2024-03-08 ff30bb6
build(project): :arrow_up: upgrade dependencies and using typing-extension
