+++ requirements.txt
@@ -115,7 +115,7 @@ tornado==6.4 ; python_version >= "3.9" and python_version < "3.12"
 tqdm==4.66.1 ; python_version >= "3.9" and python_version < "3.12"
 traitlets==5.14.1 ; python_version >= "3.9" and python_version < "3.12"
 types-python-dateutil==2.8.19.20240106 ; python_version >= "3.9" and python_version < "3.12"
-typing-extensions==4.9.0 ; python_version >= "3.9" and python_version < "3.12"
+typing-extensions==4.10.0 ; python_version >= "3.9" and python_version < "3.12"
 tzdata==2023.4 ; python_version >= "3.9" and python_version < "3.12"
 urllib3==2.1.0 ; python_version >= "3.9" and python_version < "3.12"
 uvicorn==0.25.0 ; python_version >= "3.9" and python_version < "3.12"
