+++ tests/test_actor.py
@@ -5,7 +5,14 @@
 # GitHub   : https://github.com/SongshGeo
 # Website: https://cv.songshgeo.com/
 
-from typing import Tuple, TypeAlias
+"""
+测试行动者，包括
+1. 新建的行动者拥有正确的属性；
+2. 移动到某个斑块上
+3. 死亡
+4. 获取属性值（自己或所在斑块）
+5. 设置属性值（自己或所在斑块）
+"""
 
 import pytest
 
@@ -13,40 +20,99 @@ from abses import MainModel
 from abses.actor import Actor
 from abses.cells import PatchCell
 from abses.nature import PatchModule
-from abses.sequences import ActorsList
 
 
 class TestActor:
     """Test the Actor class."""
 
-    def test_actor_attributes(self):
+    def test_attributes(self, model: MainModel):
         """测试主体的属性"""
-        model = MainModel()
-        actor = Actor(model=model)
-        layer = PatchModule.from_resolution(model)
+        # arrange / act
+        actor = model.agents.create(Actor, singleton=True)
 
+        # act / assert
         assert actor.on_earth is False
         assert actor.breed == "Actor"
-        pos = (3, 3)
-        actor.move.to(layer=layer, pos=pos)
+        assert actor.at is None
+
+    def test_movements(
+        self, model: MainModel, module: PatchModule, cell_0_0: PatchCell
+    ):
+        """Test moving"""
+        # arrange
+        actor = model.agents.create(Actor, singleton=True)
+        # act
+        pos = (0, 0)
+        actor.move.to(layer=module, pos=pos)
+        # assert
         assert actor.on_earth is True
-        assert len(actor.here) == 1
-        assert actor.here == ActorsList(model, [actor])
+        assert len(actor.at.agents) == 1
+        assert actor.at is cell_0_0
+        assert actor.layer is module
 
-    # def test_actor_selecting(self, model, farmer_cls):
-    #     """测试主体的选取"""
-    #     actor = Actor(model=model)
-    #     actor.test1 = 1
-    #     actor.test2 = "testing"
-    #     selection = {"test1": 1, "test2": "testing"}
-    #     selection2 = "test1 == 1, test2 == testing"
-    #     selection3 = "Actor"
+    def test_die(self, model: MainModel, cell_0_0: PatchCell):
+        """Test die"""
+        # arrange
+        actor1 = model.agents.create(Actor, num=1, singleton=True)
+        actor2 = cell_0_0.agents.create(Actor, singleton=True)
+        # act
+        actor1.die()
+        actor2.die()
+        # assert
+        assert actor1 not in model.agents
+        assert actor2 not in model.agents
+        assert len(model.agents) == 0
 
-    #     assert actor.selecting(selection=selection)
-    #     assert actor.selecting(selection=selection2)
-    #     assert actor.selecting(selection=selection3)
+    @pytest.mark.parametrize(
+        "attr, target, expected",
+        [
+            ("test2", None, 3),
+            ("test2", "self", 3),
+            ("test2", "at", 2),
+        ],
+    )
+    def test_get(self, cell_0_0: PatchCell, attr, target, expected):
+        """Test getting values."""
+        # arrange
+        actor = cell_0_0.agents.create(Actor, singleton=True)
+        cell_0_0.test1 = 1
+        cell_0_0.test2 = 2
+        actor.test2 = 3
+        # act
+        value = actor.get(attr=attr, target=target)
+        # assert
+        assert value == expected
 
-    #     actor2 = farmer_cls(model=model)
-    #     assert actor2.selecting(selection=selection) is False
-    #     assert actor2.selecting(selection=selection2) is False
-    #     assert actor2.selecting(selection=selection3) is False
+    @pytest.mark.parametrize(
[…]
