+++ abses/cells.py
@@ -120,7 +120,7 @@ class PatchCell(mg.Cell, _LinkNode):
         """
         return bool(self._agents) if breed is None else breed in self._agents
 
-    def get(self, attr_name: str) -> Any:
+    def get(self, attr: str) -> Any:
         """Gets the value of an attribute or registered property. Automatically update the value if it is the dynamic variable of the layer.
 
         Parameters:
@@ -135,11 +135,11 @@ class PatchCell(mg.Cell, _LinkNode):
             AttributeError:
                 Attribute value of the associated patch cell.
         """
-        if attr_name in self.layer.dynamic_variables:
-            self.layer.dynamic_var(attr_name=attr_name)
-        if not hasattr(self, attr_name):
-            raise AttributeError(f"{attr_name} not exists in {self.layer}.")
-        return getattr(self, attr_name)
+        if attr in self.layer.dynamic_variables:
+            self.layer.dynamic_var(attr_name=attr)
+        if not hasattr(self, attr):
+            raise AttributeError(f"{attr} not exists in {self.layer}.")
+        return getattr(self, attr)
 
     def linked_attr(
         self,
