### 2024-03-09 5914c68
refactor(src): :recycle: refactoring nature and its tests
