+++ abses/nature.py
@@ -7,8 +7,9 @@
 
 from __future__ import annotations
 
+import functools
 from numbers import Number
-from typing import TYPE_CHECKING, Any, Iterable, List, Optional, Self
+from typing import TYPE_CHECKING, Any, Iterable, Optional, Self
 
 import geopandas
 import mesa_geo as mg
@@ -25,6 +26,7 @@ from shapely import Geometry
 
 from abses.links import _LinkNode
 from abses.modules import CompositeModule, Module
+from abses.random import ListRandom
 
 from .actor import Actor
 from .cells import PatchCell
@@ -71,9 +73,8 @@ class PatchModule(Module, mg.RasterLayer):
         logger.info(f"Using rioxarray version: {rioxarray.__version__}")
 
         for cell in self:
-            # for cell in self.array_cells.flatten():
             cell.layer = self
-        self._file = None
+        self._updated_ticks = []
 
     @property
     def cell_properties(self) -> set[str]:
@@ -84,11 +85,6 @@ class PatchModule(Module, mg.RasterLayer):
     def attributes(self) -> set[str]:
         """All accessible attributes from this layer."""
         return self._attributes | self.cell_properties
-
-    @property
-    def file(self) -> str | None:
-        """If the module is created by reading a raster dataset, save the file path."""
-        return self._file
 
     @property
     def shape2d(self) -> Coordinate:
@@ -245,7 +241,6 @@ class PatchModule(Module, mg.RasterLayer):
             cell_cls=cell_cls,
         )
         obj._transform = dataset.transform
-        obj._file = raster_file
         obj.apply_raster(values, attr_name=attr_name)
         return obj
 
@@ -263,7 +258,7 @@ class PatchModule(Module, mg.RasterLayer):
             return self.get_raster(data)
         raise TypeError("Invalid data type or shape.")
 
-    def get_raster(self, attr_name: str | None = None) -> np.ndarray:
+    def get_raster(self, attr_name: Optional[str] = None) -> np.ndarray:
         """Obtaining the Raster layer by attribute.
 
         Parameters:
@@ -287,12 +282,15 @@ class PatchModule(Module, mg.RasterLayer):
         Returns:
             2D numpy.ndarray data of the variable.
         """
+        if self.time.tick in self._updated_ticks:
+            return super().dynamic_var(attr_name)
         array = super().dynamic_var(attr_name)
         # 判断算出来的是一个符合形状的矩阵
         self._attr_or_array(array)
         # 将矩阵转换为三维，并更新空间数据
         array_3d = array.reshape(self.shape3d)
         self.apply_raster(array_3d, attr_name=attr_name)
+        self._updated_ticks.append(self.time.tick)
         return array
 
     def get_rasterio(self, attr_name: str | None = None) -> rio.MemoryFile:
@@ -305,9 +303,12 @@ class PatchModule(Module, mg.RasterLayer):
         Returns:
             The rasterio tmp memory file of raster.
         """
-        data = self.get_raster(attr_name=attr_name)
+        if attr_name is None:
+            data = np.ones(self.shape2d)
+        else:
+            data = self.get_raster(attr_name=attr_name)
         # 如果获取到的是2维，重整为3维
-        if len(data.shape):
+        if len(data.shape) != 3:
             data = data.reshape(self.shape3d)
         with rio.MemoryFile() as mem_file:
             with mem_file.open(
@@ -322,39 +323,6 @@ class PatchModule(Module, mg.RasterLayer):
                 dataset.write(data)
             # Open the dataset again for reading and return
             return mem_file.open()
-
-    def geometric_cells(
-        self, geometry: Geometry, refer_layer: str | None = None, **kwargs
-    ) -> List[PatchCell]:
[…]
