+++ tests/test_cells.py
@@ -5,72 +5,99 @@
 # GitHub   : https://github.com/SongshGeo
 # Website: https://cv.songshgeo.com/
 
+"""测试单元格。
+1. 测试单元格能够选取它的邻居
+2. 测试单元格的属性和方法
+"""
+
 from __future__ import annotations
 
 import numpy as np
 import pytest
 
 from abses import ActorsList
+from abses.actor import Actor
 from abses.main import MainModel
 from abses.nature import PatchModule
 from abses.tools.func import get_buffer
 
 
-@pytest.fixture(name="cells")
-def cells_list():
-    """模拟一个斑块模块，包含5*5个格子"""
-    model = MainModel()
-    module = PatchModule.from_resolution(model, shape=(5, 5))
-    return model, module.array_cells
+class TestCellNeighboring:
+    """测试单元格的邻居选取问题"""
 
+    @pytest.fixture(name="array_cells")
+    def cells_list(self, model: MainModel):
+        """模拟一个斑块模块，包含5*5个格子"""
+        module = PatchModule.from_resolution(model, shape=(5, 5))
+        return module.array_cells
 
-def get_cells(array_cells, centre, radius, moor, annular, include_center):
-    """获取合理的斑块"""
-    zeros = np.zeros(shape=array_cells.shape)
-    zeros[*centre] = 1
-    arr = get_buffer(zeros, radius=radius, moor=moor, annular=annular)
-    arr[*centre] = include_center
-    return arr
+    @staticmethod
+    def get_cells(array, centre, radius, moor, annular, include_center):
+        """获取合理的数组"""
+        zeros = np.zeros(shape=array.shape)
+        zeros[*centre] = 1
+        mask = get_buffer(zeros, radius=radius, moor=moor, annular=annular)
+        mask[*centre] = include_center
+        return mask
 
-
-@pytest.mark.parametrize(
-    "centre, moore, radius, include_center, annular",
-    [
-        # Happy path tests
-        ([2, 2], False, 1, False, False),  # Test case 1
-        ([2, 2], True, 2, True, False),  # Test case 2
-        ([2, 2], False, 3, True, True),  # Test case 3
-        ([2, 2], True, 1, True, False),  # Test case 5
-        ([2, 2], False, 2, True, True),  # Test case 6
-        ([2, 2], True, 1, False, True),  # Test case 8
-        ([2, 2], False, 2, False, True),  # Test case 9
-    ],
-    ids=[
-        "Happy path - Test case 1",
-        "Happy path - Test case 2",
-        "Happy path - Test case 3",
-        # "Edge case - Test case 4",
-        "Edge case - Test case 5",
-        "Edge case - Test case 6",
-        # "Error case - Test case 7",
-        "Error case - Test case 8",
-        "Error case - Test case 9",
-    ],
-)
-def test_get_neighboring_cells(
-    cells, centre, moore, radius, include_center, annular
-):
-    """测试通过斑块搜索周围格子"""
-    # Arrange
-    model, array_cells = cells
-    expected_mask = get_cells(
-        array_cells, centre, radius, moore, annular, include_center
+    @pytest.mark.parametrize(
+        "centre, moore, radius, include_center, annular",
+        [
+            # Happy path tests
+            ([2, 2], False, 1, False, False),  # Test case 1
+            ([2, 2], True, 2, True, False),  # Test case 2
+            ([2, 2], False, 3, True, True),  # Test case 3
+            ([2, 2], True, 1, True, False),  # Test case 5
+            ([2, 2], False, 2, True, True),  # Test case 6
+            ([2, 2], True, 1, False, True),  # Test case 8
+            ([2, 2], False, 2, False, True),  # Test case 9
+        ],
+        ids=[
+            "Happy path - Test case 1",
+            "Happy path - Test case 2",
[…]
