+++ tests/test_links.py
@@ -5,11 +5,15 @@
 # GitHub   : https://github.com/SongshGeo
 # Website: https://cv.songshgeo.com/
 
+"""测试链接
+"""
+
 from typing import List
 
 import pytest
 
 from abses import Actor
+from abses.errors import ABSESpyError
 from abses.links import _LinkContainer, _LinkNode
 from abses.main import MainModel
 
@@ -143,3 +147,153 @@ class TestNetworkx:
         # assert
         assert set(graph.nodes) == set(tres_nodes)
         assert graph.number_of_edges() == 2
+
+
+class TestLinkProxy:
+    """Test linking methods in proxy."""
+
+    @pytest.mark.parametrize(
+        "links, mutual, to_or_by, expected",
+        [
+            (("test", "test1", "test2"), True, "to", (True, True)),
+            (("test", "test1", "test2"), False, "to", (True, False)),
+            (("test", "test", "test2"), False, "to", (True, False)),
+            (("test", "test1", "test1"), True, "to", (True, True)),
+            (("test", "test1", "test2"), True, "by", (True, True)),
+            (("test", "test1", "test2"), False, "by", (False, True)),
+            (("test", "test", "test2"), False, "by", (False, True)),
+            (("test", "test1", "test1"), True, "by", (True, True)),
+        ],
+    )
+    def test_link_to_or_by(
+        self,
+        tres_nodes: List[Actor],
+        links,
+        mutual,
+        to_or_by,
+        container: _LinkContainer,
+        expected,
+    ):
+        """Test linking to"""
+        # arrange
+        a1, a2, a3 = tres_nodes
+        link1, link2, link3 = links
+
+        # act
+        getattr(a1.link, to_or_by)(a2, link1, mutual=mutual)
+        getattr(a1.link, to_or_by)(a3, link2, mutual=mutual)
+        getattr(a2.link, to_or_by)(a3, link3, mutual=mutual)
+
+        # assert
+        assert a1.link == set([link1, link2])
+        assert a2.link == set([link1, link3])
+        assert a3.link == set([link2, link3])
+        assert set(container.links) == set([link1, link2, link3])
+        assert container.has_link(link1, a1, a2) == expected
+        assert container.has_link(link2, a1, a3) == expected
+        assert container.has_link(link3, a2, a3) == expected
+
+    @pytest.mark.parametrize(
+        "mutual, to_or_by, expected",
+        [
+            (True, "to", (True, True)),
+            (False, "to", (True, False)),
+            (True, "by", (True, True)),
+            (False, "by", (False, True)),
+        ],
+    )
+    def test_has(self, tres_nodes: List[Actor], mutual, to_or_by, expected):
+        """Test that if a node has a link, or has link with another node."""
+        # arrange
+        node1, node2, _ = tres_nodes
+        getattr(node1.link, to_or_by)(node2, "test", mutual=mutual)
+
+        # act / assert
+        assert node1.link.has("test", node2) == expected
+        assert node1.link.has("test") == expected
+
+    @pytest.mark.parametrize(
+        "to_or_by, mutuals, link_name, direction, expected",
+        [
+            (("to", "to"), (True, True), "t2", "out", (True, False)),
+            (("by", "to"), (False, True), "t2", "out", (False, False)),
+            (("by", "to"), (True, True), "t2", "out", (True, False)),
+            (("by", "to"), (False, True), "t2", "in", (True, False)),
+            (("by", "to"), (False, True), ["t2", "t3"], "in", (True, True)),
+            (("by", "to"), (False, True), ["t2", "t3"], "out", (False, True)),
+            (("by", "to"), (True, True), ["t2", "t3"], "in", (True, True)),
+            (("by", "to"), (True, True), ["t2", "t3"], "out", (True, True)),
+        ],
+    )
[…]
