### 2024-03-09 8470c2d
test(tests): :white_check_mark: refacatoring forest fire example codes
