### 2024-03-11 868265d
refactor(tests): :recycle: refactoring sheep-wolf model tests
