### 2024-03-11 0d4dd37
refactor(src): :recycle: using default schedule and data collector, but compatible to attrs config
