+++ abses/time.py
@@ -194,6 +194,7 @@ class TimeDriver(_Component):
                 self.go(ticks=1, **kwargs)
             return
         # tick = 1
+        self._model.schedule.step()
         self._tick += ticks
         if self.ticking_mode == "duration":
             self.dt += self.duration
