+++ abses/cells.py
@@ -12,7 +12,7 @@
 from __future__ import annotations
 
 from functools import cached_property
-from typing import TYPE_CHECKING, Any, Optional
+from typing import TYPE_CHECKING, Any
 
 import mesa_geo as mg
 

[…]
