+++ docs/api/container.md
@@ -1,4 +1,6 @@
 
 # Actors' Container
 
-:::abses.container.AgentsContainer
+:::abses.container._AgentsContainer
+
+:::abses.container._CellAgentsContainer
