+++ docs/api/human.md
@@ -7,6 +7,7 @@ date: 2023-01-10
 ## Major human module
 
 :::abses.human.BaseHuman
+
 ## Human submodule
 
 :::abses.human.HumanModule
