+++ abses/links.py
@@ -6,12 +6,27 @@
 # Website: https://cv.songshgeo.com/
 
 """主体、斑块之间可以产生连接。
+
+Actor, PatchCell can be used to create links.
 """
+
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Dict, Iterable, List, Optional, Set, Tuple
+import contextlib
+from typing import (
+    TYPE_CHECKING,
+    Dict,
+    Iterable,
+    List,
+    Literal,
+    Optional,
+    Set,
+    Tuple,
+)
 
+with contextlib.suppress(ImportError):
+    import networkx as nx
 try:
     from typing import TypeAlias
 except ImportError:
@@ -25,7 +40,8 @@ if TYPE_CHECKING:
     from abses.actor import Actor
     from abses.cells import PatchCell
 
-Node: TypeAlias = "Actor | PatchCell"
+LinkingNode: TypeAlias = "Actor | PatchCell"
+Direction: TypeAlias = Optional[Literal["in", "out"]]
 
 
 class _LinkContainer:
@@ -48,7 +64,7 @@ class _LinkContainer:
         self._back_links[link_name] = {}
 
     def owns_links(
-        self, node: _LinkNode, direction: Optional[str] = "out"
+        self, node: LinkingNode, direction: Direction = "out"
     ) -> Tuple[str]:
         """The links a specific node owns."""
         if direction == "out":
@@ -64,20 +80,27 @@ class _LinkContainer:
         links = {link for link, agents in data.items() if node in agents}
         return tuple(links)
 
-    def get_graph(self, link_name):
-        """Get the graph."""
-        try:
-            import networkx as nx
-        except ImportError as exc:
+    def get_graph(self, link_name: str) -> "nx.Graph":  # type: ignore
+        """Get the networkx graph.
+
+        Parameters:
+            link_name:
+                The link name for converting into a graph.
+
+        Raises:
+            ImportError:
+                If networkx is not installed.
+        """
+        if "nx" not in globals():
             raise ImportError(
                 "You need to install networkx to use this function."
-            ) from exc
+            )
         graph = nx.from_dict_of_lists(self._links[link_name])
         self._cached_networks[link_name] = graph
         return graph
 
     def _register_link(
-        self, link_name: str, source: _LinkNode, target: _LinkNode
+        self, link_name: str, source: LinkingNode, target: LinkingNode
     ) -> None:
         """Register a link."""
         if link_name not in self._links:
@@ -88,26 +111,54 @@ class _LinkContainer:
             self._back_links[link_name][target] = set()
 
     def has_link(
-        self, link_name: str, node1: _LinkNode, node2: _LinkNode
+        self, link_name: str, source: LinkingNode, target: LinkingNode
     ) -> tuple[bool]:
-        """If link exists."""
+        """If any link exists between source and target.
+
+        Parameters:
+            link_name:
+                The name of the link.
+            source:
+                The source node.
+            target:
+                The target node.
+
+        Raises:
+            KeyError:
+                If the link name does not exist.
+
+        Returns:
+            tuple:
+                A tuple of two booleans.
+                The first element is True if the link exists from source to target.
+                The second element is True if the link exists from target to source.
+        """
+        if link_name not in self.links:
+            raise KeyError("No link named {link_name}.")
         data = self._links[link_name]
-        node1_to_node2 = (
-            False if node1 not in data else node2 in data.get(node1, [])
-        )
-        node2_to_node1 = (
-            False if node2 not in data else node1 in data.get(node2, [])
-        )
-        return node1_to_node2, node2_to_node1
+        to = False if source not in data else target in data.get(source, [])
+        by = False if target not in data else source in data.get(target, [])
[…]
