### 2024-03-12 af72286
docs(docs): :recycle: refactoring the structure of api docs
