+++ docs/api/links.md
@@ -1,0 +1,7 @@
+---
+title: Agents' Interlinks
+authors: SongshGeo
+date: 2024-03-13
+---
+
+:::abses.links._LinkProxy
