### 2024-03-12 23b3210
docs(docs): :memo: updating beginner docs notebooks
