+++ abses/move.py
@@ -11,7 +11,12 @@ This script is used to manipulate actors' movements.
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Literal, Optional, Tuple, TypeAlias
+from typing import TYPE_CHECKING, Any, Literal, Optional, Tuple
+
+try:
+    from typing import TypeAlias
+except ImportError:
+    from typing_extensions import TypeAlias
 
 import mesa_geo as mg
 from mesa.space import Coordinate
