### 2024-03-12 9c7ed3f
docs(project): :bookmark: change log update for version 0.5.0
