### 2024-03-12 ABSESpy/ABSESpy#37 @SongshGeo
@jamesdamillington Sorry for keeping you waiting. I'm glad to notify you that this problem is incompatible with Python 3.9 and should be fixed in the newest PR #41. In addition to fixing the bug, many improvements were made. I hope you enjoy using `abses`!
