+++ tests/test_sheep_wolf.py
@@ -1,0 +1,3 @@
+import pytest
+
+from .examples.sheep_wolf import WolfSheepModel
