### 2024-03-19 ABSESpy/ABSESpy#37 @jamesdamillington
@SongshGeo I tried removing my `conda env` I created previously and then reinstalled `abses` for python `version=3.9`. Unfortunately, I still get the same 'typing' error as before. However, I am not convinced I properly removed and reinstalled as many dependencies did not download because apparently they were already satisfied by packages in an environment with the same name I thought I had removed...

So I created a `conda env` with `python=3.10` and did `pip install abses`. With that the install seems to work¬

But now I have other issues (maybe they should be raised in a new issue?):
- running `model.py` returns an error:  `NameError: name 'Human' is not defined` from the line `model=MainModel(human_class=Human, nature_class=Nature)` I seem to be able to fix that by changing the line to `model=MainModel(human_class=CustomHuman, nature_class=CustomNature)`
- when I try to run the [Forest Fire](https://absespy.github.io/ABSESpy/tutorial/beginner/fire_tutorial/) test I get the error `ModuleNotFoundError: No module named 'tests'`
