### 2024-03-20 ABSESpy/ABSESpy#37 @SongshGeo
@jamesdamillington

I appriciate your feedbacks. I'm so sorry to hear that Python 3.9 still doesn't work. I created a new env with `poetry` and `pyenv` (I haven't used `conda` for a year), and a Python 3.9 version works well from my side.  I will do more tests with `conda` env and different platforms. It takes time. I will inform you of any updates.

I believe the documentation and tutorials regarding the other two issues have not been updated with the newest API. I created a new issue #43 and will go through them as soon as possible. :)
