### 2024-03-20 8b08765
docs(tests): :white_check_mark: re-wrote hotelling law model for newest API
