+++ examples/hotelling_law/model.py
@@ -1,16 +1,123 @@
+#!/usr/bin/env python 3.11.0
+# -*-coding:utf-8 -*-
+# @Author  : Shuang (Twist) Song
+# @Contact   : SongshGeo@gmail.com
+# GitHub   : https://github.com/SongshGeo
+# Website: https://cv.songshgeo.com/
+
+"""
+Hotelling's Law model.
+"""
+
 import numpy as np
+from mesa.time import SimultaneousActivation
+from scipy.spatial.distance import cdist
 
-from abses import MainModel
-
-from .shop import Shop
-
-# Instantiate a random number generator
-rng = np.random.default_rng()
+from abses import Actor, ActorsList, MainModel, PatchCell
 
 
-# Calculate euclidean distance between two points
-def euclidean_distance(x1, y1, x2, y2):
-    return np.sqrt((x1 - x2) ** 2 + (y1 - y2) ** 2)
+class Customer(PatchCell):
+    """
+    Each patch cell represents a customer.
+    Customer prefers to buy from the nearest & cheapest shop.
+    """
+
+    def find_preference(self):
+        """Find the nearest & cheapest shop."""
+        stores: ActorsList[Actor] = self.model.actors
+        # Create a list of all shops
+        prices = stores.array("price")
+        # Create a list of all distances from the customer to each shop
+        distances = cdist(
+            np.array([self.pos]),
+            np.array([shop.at.pos for shop in stores]),
+        )[0]
+        # Pair each shop to its distance & price
+        _pair = dict(zip(stores, distances + prices))
+        prefer_store = min(_pair, key=_pair.get)
+        self.link.by(prefer_store, link_name="prefer")
+
+
+class Shop(Actor):
+    """
+    Shop agent
+    """
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.price = 10
+        self.next_position = None
+        self.next_price = None
+
+    @property
+    def area_count(self) -> int:
+        """Return the number of customers in the shop's area."""
+        return len(self.link.get("prefer", direction="out"))
+
+    def step(self):
+        self.adjust_price()
+        self.adjust_position()
+
+    def advance(self):
+        self.affect_price()
+        self.affect_position()
+
+    def adjust_price(self):
+        """Evaluate the potential revenue for each possible price change.
+        Choose the one with the highest potential revenue."""
+        # Save initial price
+        init_price = self.price
+
+        # Build a list of all possible prices
+        _possible_prices = [init_price - 1, init_price, init_price + 1]
+
+        # Pair each possible price change to its potential revenue
+        _potential_revenues = {}
+        for price in _possible_prices:
+            self.price = price
+            self.model.recalculate_preferences()
+            _potential_revenues[price] = self.area_count * price
+        # Check if all potential revenues are 0
+        # if so, decrease price by 1
+        if all(value == 0 for value in _potential_revenues.values()):
+            self.next_price = self.price - 1
+        # Otherwise, choose the price with the highest potential revenue
+        else:
+            self.next_price = max(
+                _potential_revenues, key=_potential_revenues.get
+            )
+
+        # Reset price to initial price
+        self.price = init_price
+
+    def adjust_position(self):
+        """Evaluate the potential areas for each possible move.
+        Choose the one with the highest potential area."""
+        cell_now = self.at
+        # Get all possible candidates for the next position
+        _possible_moves = self.at.neighboring(moore=True, include_center=False)
+
+        # Pair each possible move to their potential areas
+        _potential_areas = {}
+        for move in _possible_moves:
+            self.move.to(move)
+            self.model.recalculate_preferences()
+            _potential_areas[move] = self.area_count
+
[…]
