### 2024-03-20 e832741
style(project): :truck: seperate api and examples
