### 2024-03-20 dab3084
style(tests): :truck: grouping tests and configs
