+++ docs/examples/official.md
@@ -20,6 +20,13 @@ This showcases the use of `ABSESpy` for ABM. We divide the models into two types
 
     <!-- [:octicons-arrow-right-24: Checkout] -->
 
+-   :material-robot-happy:{ .lg .middle } __[Predator-Prey]__
+
+    ---
+    An implementation of the Wolf-sheep ecosystem model.
+
+    <!-- [:octicons-arrow-right-24: Checkout] -->
+
 </div>
 
 
@@ -50,6 +57,7 @@ This showcases the use of `ABSESpy` for ABM. We divide the models into two types
 <!-- Heuristic models -->
   [Hotelling]: ../tutorial/beginner/hotelling_tutorial.ipynb
   [Forest Fire]: ../tutorial/beginner/fire_tutorial.ipynb
+  [Predator-Prey]: ../tutorial/beginner/predation_tutorial.ipynb
 
 <!-- Real-world models -->
   [South China Livelihood]: official/south_china_livelihood.md
