### 2024-03-20 beec2f9
docs(docs): :memo: make sure tutorials notebooks are runable
