+++ docs/tutorial/tutorial.md
@@ -50,10 +50,10 @@ You have gained a thorough understanding of the agent-based model. Now, you are 
 
 <div class="grid" markdown>
 
-[:material-book: __A01__](advanced/human_behavior_simulation.ipynb) How to apply an advanced [human behavior simulation framework]?
+[:material-book: __A01__](advanced/geodata.ipynb) Including real-world geographic datasets.
 { .card }
 
-[:material-book: __A02__](advanced/geodata.ipynb) Including real-world geographic datasets.
+[:material-book: __A02__](advanced/human_behavior_simulation.ipynb) Implement an advanced [human behavior simulation framework] automatically.
 { .card }
 
 > :material-book: __Huh__ ... More tutorials in this level are coming soon.
