+++ docs/examples/official/south_china_livelihood.md
@@ -1,5 +1,28 @@
+---
+title: South China Livelihood Evolution
+author: Shuang Song
+date: 2024-03-20
+---
+
 !!! Warning
 
-    The document is under active development, and related papers are also being produced. Please [contact the author] for any related inquiries.
+    This is a finished model with a in-production paper. Once the paper is accepted, source codes would be fully released. Please [contact the author] for any related inquiries.
+
+## Overview
+
+This project's source code is available at this [GitHub repo].
+
+This project constructs an agent-based model that simulates the spread of agriculture and the interaction between farmers (invaders) and hunter-gatherers (natives) in southern China. Based on three different kinds of subsistence economies (dryland farming, paddy farming, hunting and gathering), the model focuses on the cultural lifestyles of human groups, involving mobility, population limitations and growth rates, the possibility of diffusion and conversion, and competitive strength. Groups of farmers enter the map alongside native groups of hunter-gatherers and initiate their own development, possible conversion, and encountered conflict.
+
+Here is the overall workflow of the model:
+
+![fig](https://songshgeo-picgo-1302043007.cos.ap-beijing.myqcloud.com/uPic/fig.png)
+
+## Result glance
+
+The simulation results reveal the sluggish development of farmers under the complex interaction between different cultural groups. This is related to the abundant natural environment of southern China, which acts as a friction zone for the spread of agriculture. The results also provide evidence that once farming accumulates enough advantages in a suitable environment, its development will become rapid and unstoppable.
+
+![vM4vIY](https://songshgeo-picgo-1302043007.cos.ap-beijing.myqcloud.com/uPic/vM4vIY.jpg)
 
   [contact the author]: mailto:songshgeo@mail.bnu.edu.cn
+  [GitHub repo]: https://github.com/SongshGeo/South-China-Livehood-Evolution
