+++ README.md
@@ -1,57 +1,77 @@
+---
+title: ABSESpy
+author: Shuang Song
+---
+
 ![ABSES_banner](https://songshgeo-picgo-1302043007.cos.ap-beijing.myqcloud.com/uPic/CleanShot%202023-10-19%20at%2019.08.12@2x.png)
 
-<div style="text-align: center;">
-    <!-- License Badge -->
-    <a href="http://www.apache.org/licenses/">
-        <img src="https://img.shields.io/github/license/absespy/ABSESpy" alt="license">
+<div align="center"><p>
+    <a href="https://github.com/absespy/ABSESpy/releases/latest">
+      <img alt="Latest release" src="https://img.shields.io/github/v/release/absespy/ABSESpy?style=for-the-badge&logo=starship&color=C9CBFF&logoColor=D9E0EE&labelColor=302D41" />
     </a>
-    <!-- Downloads Badge -->
-    <img src="https://img.shields.io/github/downloads/absespy/ABSESpy/total" alt="downloads">
-    <!-- Code Size Badge -->
-    <img src="https://img.shields.io/github/languages/code-size/absespy/ABSESpy" alt="codesize">
-    <!-- Tag Badge -->
-    <img src="https://img.shields.io/github/v/tag/absespy/ABSESpy" alt="tag">
-    <br>
-    <!-- Website Badge -->
-    <a href="https://cv.songshgeo.com/">
-        <img src="https://img.shields.io/badge/Website-SongshGeo-brightgreen.svg" alt="github">
+    <a href="https://github.com/absespy/ABSESpy/pulse">
+      <img alt="Last commit" src="https://img.shields.io/github/last-commit/absespy/ABSESpy?style=for-the-badge&logo=starship&color=8bd5ca&logoColor=D9E0EE&labelColor=302D41"/>
     </a>
-    <!-- Stars Badge -->
-    <img src="https://img.shields.io/github/stars/absespy/ABSESpy?style=social" alt="stars">
-    <!-- Twitter Badge -->
-    <a href="https://twitter.com/shuangsong11">
-        <img src="https://img.shields.io/twitter/follow/shuangsong11?style=social" alt="twitter">
+    <a href="https://github.com/absespy/ABSESpy/blob/main/LICENSE">
+      <img alt="License" src="https://img.shields.io/github/license/absespy/ABSESpy?style=for-the-badge&logo=starship&color=ee999f&logoColor=D9E0EE&labelColor=302D41" />
     </a>
+    <a href="https://github.com/absespy/ABSESpy/stargazers">
+      <img alt="Stars" src="https://img.shields.io/github/stars/absespy/ABSESpy?style=for-the-badge&logo=starship&color=c69ff5&logoColor=D9E0EE&labelColor=302D41" />
+    </a>
+    <a href="https://github.com/absespy/ABSESpy/issues">
+      <img alt="Issues" src="https://img.shields.io/github/issues/absespy/ABSESpy?style=for-the-badge&logo=bilibili&color=F5E0DC&logoColor=D9E0EE&labelColor=302D41" />
+    </a>
+    <a href="https://github.com/absespy/ABSESpy">
+      <img alt="Repo Size" src="https://img.shields.io/github/repo-size/absespy/ABSESpy?color=%23DDB6F2&label=SIZE&logo=codesandbox&style=for-the-badge&logoColor=D9E0EE&labelColor=302D41" />
+    </a>
+    <!-- <a href="https://patreon.com/chrisatmachine" title="Donate to this project using Patreon">
+      <img alt="Patreon donate button" src="https://img.shields.io/badge/patreon-donate-yellow.svg?style=for-the-badge&logo=starship&color=f5a97f&logoColor=D9E0EE&labelColor=302D41" />
+    </a> -->
[…]
