### 2024-03-20 9401f3b
docs(docs): :memo: adjust max pixel for readme fig
