### 2024-03-20 ed5b78a
fix(docs): :memo: fix redundant contents in readme
