+++ README.md
@@ -4,29 +4,27 @@
 
 <div align="center"><p>
     <a href="https://github.com/absespy/ABSESpy/releases/latest">
-      <img alt="Latest release" src="https://img.shields.io/github/v/release/absespy/ABSESpy?style=for-the-badge&logo=probot&color=C9CBFF&logoColor=D9E0EE&labelColor=302D41" />
+        <img alt="Latest release" src="https://img.shields.io/github/v/release/absespy/ABSESpy?style=for-the-badge&logo=probot&color=C9CBFF&logoColor=D9E0EE&labelColor=302D41" />
     </a>
     <a href="https://github.com/absespy/ABSESpy/pulse">
-      <img alt="Last commit" src="https://img.shields.io/github/last-commit/absespy/ABSESpy?style=for-the-badge&logo=probot&color=8bd5ca&logoColor=D9E0EE&labelColor=302D41"/>
+        <img alt="Last commit" src="https://img.shields.io/github/last-commit/absespy/ABSESpy?style=for-the-badge&logo=probot&color=8bd5ca&logoColor=D9E0EE&labelColor=302D41"/>
     </a>
     <a href="https://github.com/absespy/ABSESpy/blob/main/LICENSE">
-      <img alt="License" src="https://img.shields.io/github/license/absespy/ABSESpy?style=for-the-badge&logo=probot&color=ee999f&logoColor=D9E0EE&labelColor=302D41" />
+        <img alt="License" src="https://img.shields.io/github/license/absespy/ABSESpy?style=for-the-badge&logo=probot&color=ee999f&logoColor=D9E0EE&labelColor=302D41" />
     </a>
     <a href="https://github.com/absespy/ABSESpy/stargazers">
-      <img alt="Stars" src="https://img.shields.io/github/stars/absespy/ABSESpy?style=for-the-badge&logo=probot&color=c69ff5&logoColor=D9E0EE&labelColor=302D41" />
+        <img alt="Stars" src="https://img.shields.io/github/stars/absespy/ABSESpy?style=for-the-badge&logo=probot&color=c69ff5&logoColor=D9E0EE&labelColor=302D41" />
     </a>
     <a href="https://github.com/absespy/ABSESpy/issues">
-      <img alt="Issues" src="https://img.shields.io/github/issues/absespy/ABSESpy?style=for-the-badge&logo=probot&color=F5E0DC&logoColor=D9E0EE&labelColor=302D41" />
+        <img alt="Issues" src="https://img.shields.io/github/issues/absespy/ABSESpy?style=for-the-badge&logo=probot&color=F5E0DC&logoColor=D9E0EE&labelColor=302D41" />
     </a>
     <a href="https://github.com/absespy/ABSESpy">
-      <img alt="Repo Size" src="https://img.shields.io/github/repo-size/absespy/ABSESpy?color=%23DDB6F2&label=SIZE&logo=codesandbox&style=for-the-badge&logoColor=D9E0EE&labelColor=302D41" />
+        <img alt="Repo Size" src="https://img.shields.io/github/repo-size/absespy/ABSESpy?color=%23DDB6F2&label=SIZE&logo=codesandbox&style=for-the-badge&logoColor=D9E0EE&labelColor=302D41" />
     </a>
-    <!-- <a href="https://patreon.com/chrisatmachine" title="Donate to this project using Patreon">
-      <img alt="Patreon donate button" src="https://img.shields.io/badge/patreon-donate-yellow.svg?style=for-the-badge&logo=probot&color=f5a97f&logoColor=D9E0EE&labelColor=302D41" />
-    </a> -->
[…]
