### 2024-03-20 b049e4c
docs(project): :bookmark: update to version 0.5.1
