### 2024-03-22 #6298 @jamesdamillington
I have completed my review of the software and checked off all items in my checklist satisfactorily.

Before I can check off items regarding the paper, I have comments for @SongshGeo and colleagues to address as follows:

- [ ] Line 24: 'research' not 'researches'
- [ ] Line 44: rather than 'et al.' maybe use actual words (e.g. 'and others') so as not to confuse against the file suffixes which are similar abbreviations
- [ ] Line 55: What do you mean by 'practicing' here? This doesn't seem right. Please edit to clarify
- [ ] Lines 57-65: I'm surprised these three points (Perceptions, Decision-making, Response) don't match the words used in Fig 2  (Options, Evaluate, Behaviour). Or are the latter three (in the Fig) all part of the 'decision-making' step?
Aligning the steps in the list with the figure would be useful, I think
- [ ] Line 76: I think 'vary' should be 'varying'
- [ ] Line 93: 'more accurate' - this is a relative statement, so please clarify 'more accurate' than what?
- [ ] Line 99: ( wang2022h? )) is not included in the reference list
- [ ] Line 100: it's good that you recognise the similarity here to `AgentPy` but you don't then clearly explain how `absespy` is beneficial for SES researchers - maybe you could highlight the explicit functionality for representing the 'nature' side of CHANS (`AgentPy` really focuses on the 'human' side).
- [ ] Line 108: 'merely heuristic' - I think this is a little over-critical of NetLogo, which can incorporate 'real-world' (I think you mean 'empirical'?) data although not at the scale `absespy` could. I suggest you edit here to focus on the value of `absespy` for working with large-scale, empirical data so that models can run more efficiently than would be possible for the same data in NetLogo. You might also highlight your `TimeDriver` module which is a benefit over NetLogo's more simple 'ticks'
