### 2024-03-25 9e26a34
Improved on the content of the predation tutorial model
