### 2024-03-25 b429c70
feat(src): :sparkles: getting an attr value from a ActorsList
