### 2024-03-25 ABSESpy/ABSESpy#46 @wang-boyu
docs: software names in joss paper
This is part of the [joss review process](https://github.com/openjournals/joss-reviews/issues/6298).

I've noticed a few capitalization issues with some software names in the joss paper, as listed below:

- L105 & L151: netlogo and Netlogo should be NetLogo
- L42, L98, L101 & L153 : mesa-geo and Mesa-geo should be Mesa-Geo
- L95, L97, L98, L102, L129 & L153 : mesa should be Mesa
- L96: abce should be ABCE
- L128, L148 & L154: python should be Python
