### 2024-03-26 43ede69
perf(src): :zap: improve getting performance from container
