### 2024-03-26 2356698
feat(src): :sparkles: possible to control max lenght when customize PatchCell
