+++ abses/container.py
@@ -396,6 +396,26 @@ class _AgentsContainer(dict):
         """
         return self.get().apply(func, *args, **kwargs)
 
+    def item(self, how: str = "item", index: int = 0) -> Actor | None:
+        """Retrieve one agent if possible.
+
+        Parameters:
+            how:
+                The method to use to retrieve the agent.
+                Can be either "only", "item", or "random".
+                If "only", it will return the only agent in the container.
+                In this case, the container must have only one agent.
+                If more than one or no agent is found, it will raise an error.
+                If "item", it will return the agent at the given index.
+                If "random", it will return a randomly chosen agent.
+            index:
+                The index of the agent to retrieve.
+
+        Returns:
+            The agent if found, otherwise None.
+        """
+        return self.get().item(how=how, index=index)
+
 
 class _CellAgentsContainer(_AgentsContainer):
     """Container for agents located at cells."""
