### 2024-03-26 3095105
ci(project): :construction_worker: adding auto release ci
