### 2024-03-26 88fa93f
docs(docs): :bookmark: update to version 0.5.2
