### 2024-03-26 98321e5
ci(project): :green_heart: fixing release drafter to the latest version
