### 2024-03-26 bd56f10
fix(actors): :bug: now update position attribute correctly after moving
