### 2024-03-26 505e882
fix(actors): :bug: now moving has a return to control continue move or not
