### 2024-03-26 3855c54
fix(actors): :bug: Now only alive actors can call default methods.
