+++ abses/__init__.py
@@ -25,10 +25,12 @@ __all__ = [
     "ActorsList",
     "PatchCell",
     "perception",
+    "alive_required",
+    "time_condition",
 ]
 __version__ = "v0.5.2"
 
-from .actor import Actor, perception
+from .actor import Actor, alive_required, perception
 from .container import ActorsList
 from .decision import Decision
 from .human import BaseHuman
