### 2024-03-27 2c843cd
fix(docs): :pencil2: fixing typos according to reviewers' comments
