+++ docs/paper/paper.md
@@ -32,33 +32,35 @@ bibliography: "../refs.bib"
 
 ## Statement of need
 
-Social-ecological systems (SES) represent an integrated concept that recognizes the complex and interdependent dynamics between human societies and ecological systems [@folke2010]. Consisting of decision-making agents (representing people, communities, organizations, and environmental components) capable of following heterogeneous objectives [@levin2013], SES has specific needs for research support from agent-based modeling.
+Social-ecological systems (SES) are an integrated concept that recognizes the complex and interdependent dynamics between human societies and ecological systems [@folke2010]. SES consists of decision-making agents (representing people, communities, organizations, and environmental components) capable of following heterogeneous objectives [@levin2013]. As such, SES has specific needs for research support from agent-based modeling.
 
-However, ABMs' potential is yet to be fully realized in SES researches. Current challenges, such as incorporating human decision-making, portraying socio-ecological networks, and modeling real-world systems, must be addressed [@schulze2017]. Additionally, issues related to data availability, model validation, replicability, and transparency must be systematically tackled to enhance the reliability and applicability of ABM in this field [@gotts2019].
+However, the potential of ABMs has yet to be fully realized in SES research. Current challenges, such as incorporating human decision-making, portraying socio-ecological networks, and modeling real-world systems, must be addressed [@schulze2017]. Additionally, issues related to data availability, model validation, replicability, and transparency must be systematically tackled to enhance the reliability and applicability of ABM in this field [@gotts2019].
 
 Developing and refining ABM approaches for social-ecological systems are crucial in light of these needs and challenges [@reyers2018]. At the heart of this should be a modeling framework that is portable, scale-flexible, and capable of expressing the interaction of the decision-making agent with the natural environment or ecosystem. `ABSESpy` represents a significant advancement in this regard, offering several features that address the current gaps in SES modeling.
 
 ## Design structures
 
-`ABSESpy` introduces a Branch-Leaf architecture central to its functionality. It facilitates a clear separation of the human and natural subsystems within SES research, aligning with the requisite to enhance replicability and extensibility (**Figure 1**).
+`ABSESpy` introduces a branch-leaf architecture that is central to its functionality. It facilitates a clear separation of the human and natural subsystems within SES research, aligning with the requisite to enhance replicability and extensibility (**Figure 1**).
 
 ![Structure of main components of `ABSESpy` and its Branch-Leaf architecture of modules.](structure.png)
 
-Integrated by the `MainModel`, the two primary base modules are named as `Base Human` and `Base Nature`, corresponding to components of a typical SES [@reyers2018]. By this architecture, `ABSESpy` enables the addition of specialized sub-modules, thus promoting a tailored modeling approach. The extension `mesa-geo` is embedded as the base driver for the nature subsystem so that most of the different geographic datasets are compatible (`.tif`, `.nc`, `.json`, `.shp`, et al.).
+Integrated by the `MainModel`, the two primary base modules are named `Base Human` and `Base Nature`, corresponding to components of a typical SES [@reyers2018]. By this architecture, `ABSESpy` enables the addition of specialized sub-modules, thus promoting a tailored modeling approach. The extension `Mesa-Geo` is embedded as the base driver for the nature subsystem so that most of the different geographic datasets are compatible (`.tif`, `.nc`, `.json`, `.shp`, and others.).
 
 In the SES context, `ABSESpy` conceptualizes agents as `Actors` managed within a unique `ActorsContainer` and can be referred from a temporary `ActorsList`. In human sub-modules, users can define a series of `Actor`'s references by or link each other (between agent and patch, or agent and agent) by inputting advanced query. It simplifies the agents' organization, ensuring each actor can be searched, operated, and able to access global information.
 
[…]
