### 2024-03-28 2309730
docs(docs): :memo: improve JOSS paper throughoutly
