### 2024-03-28 d607ec9
docs(docs): :memo: update references info of the JOSS paper
