+++ docs/paper/paper.md
@@ -32,19 +32,19 @@ bibliography: "../refs.bib"
 
 ## Statement of need
 
-Social-ecological systems (SES) are integrated concepts recognizing the complex and interdependent dynamics between human societies and ecological systems [@folke2010]. SES consists of decision-making actors (i.e., agents representing people, communities, organizations, and environmental components) capable of following heterogeneous objectives [@levin2013]. As such, SES has specific needs for research support from agent-based modeling (ABM).
+Social-ecological systems (SES) are integrated concepts recognizing the complex and interdependent dynamics between human societies and ecological systems [@folke2010]. SES consists of decision-making actors (i.e., agents representing people, communities, organizations, and environmental components) capable of following heterogeneous objectives (i.e., agents) [@levin2013]. As such, SES has specific needs for research support from agent-based modeling (ABM).
 
-However, since few models can effectively represent the real-world SES problem, the potential of ABMs has yet to be fully realized in SES research [@schulze2017]. One reason is the variability of actors' interaction with each other or geographic spaces, including learning, decision-making, and contagion across social networks [@reyers2018]. Another reason is the complexity of managing data, parameters, and social and natural processes, which makes it hard to maintain large projects [@davidson2024]. Therefore, facilitating SES research with ABM requires novel tools to implement the interactions between actors and real-world-like natural processes with manageability and reproducibility.
+However, since few models can effectively represent the real-world SES problem, the potential of ABMs has yet to be fully realized in SES research [@schulze2017]. One reason is the variability of actors' interaction with each other or geographic spaces, including learning, decision-making, and contagion across social networks [@reyers2018]. Another reason is the complexity of managing data, parameters, and social and natural processes [@davidson2024], which makes it hard to maintain large projects. Therefore, facilitating SES research with ABM requires novel tools [@schluter2023]to implement the interactions between actors and real-world-like natural processes with manageability and reproducibility.
 
-The popular programming language `Python` is among the first choices for scientists who care about ABM, and `Mesa` is the most widely used implementation framework. It remains extensible for different domains while providing the core functionality of ABM. Therefore, customizing `Mesa` by enhancing real-world representation could be a good start for specializing in SES research. It includes expressing decision-making actors' interactions, social-ecological networks, calendar time, raster data, and extensible sub-systems. To this end, `ABSESpy` is a `Mesa` package and an advanced framework that fills gaps in SES modeling with these features.
+The popular programming language `Python` is among the first choices for scientists who care about ABM, and `Mesa` [@kazil2020a] is the most widely used implementation framework. It remains extensible for different domains while providing the core functionality of ABM. Therefore, customizing `Mesa` by enhancing real-world representation could be a good start for specializing in SES research. It includes expressing decision-making actors' interactions, social-ecological networks, calendar time, raster data, and extensible sub-systems. To this end, `ABSESpy` is a `Mesa` package and an advanced framework that fills gaps in SES modeling with these features.
 
 ## Design structures
 
 `ABSESpy`'s low-coupling design enables the implementation of maintainable projects by separating the human and natural subsystems within SESs (**Figure 1**). While the human subsystem manages actors (i.e., agents) and their interactions, the natural submodule typically handles how actors live in, move through, and interact with grid-like patch layers. This architecture also supports adding specialized submodules in each subsystem to represent varying processes with modularity.
 
[…]
