+++ docs/refs.bib
@@ -399,3 +399,44 @@
   doi = {10.1177/1464993413486544},
   language = {english}
 }
+
+@article{davidson2024,
+  title = {Simulating institutional heterogeneity in sustainability science},
+  author = {Davidson, Michael R. and Filatova, Tatiana and Peng, Wei and Verbeek, Liz and Kucuksayacigil, Fikri},
+  year = {2024},
+  month = feb,
+  journal = {Proceedings of the National Academy of Sciences},
+  volume = {121},
+  number = {8},
+  pages = {e2215674121},
+  issn = {0027-8424, 1091-6490},
+  doi = {10.1073/pnas.2215674121},
+  urldate = {2024-02-23},
+  language = {en}
+}
+
+@article{schluter2023,
+  title = {Unraveling complex causal processes that affect sustainability requires more integration between empirical and modeling approaches},
+  author = {Schlüter, Maja and Brelsford, Christa and Ferraro, Paul J. and Orach, Kirill and Qiu, Minghao and Smith, Martin D.},
+  date = {2023-10-10},
+  journaltitle = {Proceedings of the National Academy of Sciences},
+  shortjournal = {Proc. Natl. Acad. Sci. U.S.A.},
+  volume = {120},
+  number = {41},
+  pages = {e2215676120},
+  issn = {0027-8424, 1091-6490},
+  doi = {10.1073/pnas.2215676120},
+  language = {en}
+}
+
+@article{hoyer2017xarray,
+  title     = {xarray: {N-D} labeled arrays and datasets in {Python}},
+  author    = {Hoyer, S. and J. Hamman},
+  journal   = {Journal of Open Research Software},
+  volume    = {5},
+  number    = {1},
+  year      = {2017},
+  publisher = {Ubiquity Press},
+  doi       = {10.5334/jors.148},
+  url       = {https://doi.org/10.5334/jors.148}
+}
