### 2024-03-28 944a039
docs(project): :bookmark: update to version 0.5.4
