### 2024-03-28 #6298 @SongshGeo
@jamesdamillington Thanks for the valuable comments! I followed your checklist of comments one by one. Besides the corrections, here are three responses:

...
- [x] Regarding the misalignment of three points with the figure. I'm sorry about the previous misleading figure. I have now updated Figure 2. Perception, decision-making, and response are three steps in the workflow, different from the "options, evaluate, action" (three points for decision-making).
...
- [x] Your suggestion regarding comparing `AgentPy` and `NetLogo` was insightful. I updated the whole paper to explain better `ABSESpy`'s advantages. I hope the new version makes our contributions more clear:

> "The design focuses on two core improvements: (1) handling actors and their interactions and (2) enhancing reality and manageability of ABMs."

@wang-boyu  I also followed your suggestions and corrected the names. So, I closed the issue. Did I miss anything?

@crvernon, I sincerely thank you and the two reviewers for your time, and I hope you are satisfied with my responses and the latest significant revisions. I look forward to publishing the paper on JOSS.
