+++ api/human/human.md
@@ -1,0 +1,13 @@
+---
+title: human
+authors: SongshGeo
+date: 2023-01-10
+---
+
+## Major human module
+
+:::abses.human.BaseHuman
+
+## Human submodule
+
+:::abses.human.HumanModule

[…]
