### 2024-03-28 ABSESpy/ABSESpy#46 @wang-boyu
Thanks! The names look correct in the text. However, you might still need to fix those in the references section (sorry):

- L135 & L163 `mesa` --> `Mesa`
- L161 `Netlogo` --> `NetLogo`
- L163 `Mesa-geo` --> `Mesa-Geo`

I think these came from the `refs.bib` file. In this file, you might need to wrap these names with curly braces to preserve capitalization. Otherwise bibtex seems to do some case conversion automatically (see https://tex.stackexchange.com/a/10775).
