### 2024-03-29 e96d140
fix(docs): :pencil2: fixing capitalizing letters of software names in the references
