+++ docs/refs.bib
@@ -233,7 +233,7 @@
 }
 
 @inproceedings{kazil2020a,
-  title = {Utilizing Python for Agent-Based Modeling: The Mesa Framework},
+  title = {Utilizing Python for Agent-Based Modeling: The {Mesa} Framework},
   shorttitle = {Utilizing Python for Agent-Based Modeling},
   booktitle = {Social, Cultural, and Behavioral Modeling},
   author = {Kazil, Jackie and Masad, David and Crooks, Andrew},
@@ -351,7 +351,7 @@
 }
 
 @inproceedings{tisue2004netlogo,
-  title = {Netlogo: A simple environment for modeling complexity},
+  title = {{NetLogo}: A simple environment for modeling complexity},
   booktitle = {International conference on complex systems},
   author = {Tisue, Seth and Wilensky, Uri},
   year = {2004},
@@ -372,7 +372,7 @@
 }
 
 @inproceedings{wang_mesa-geo_2022,
-  title = {Mesa-geo: A GIS extension for the mesa agent-based modeling framework in python},
+  title = {{Mesa-Geo}: A GIS extension for the {Mesa} agent-based modeling framework in python},
   booktitle = {Proceedings of the 5th ACM SIGSPATIAL international workshop on GeoSpatial simulation},
   author = {Wang, Boyu and Hess, Vincent and Crooks, Andrew},
   year = {2022},
