### 2024-03-29 d821744
fix(docs): :pencil2: fixing capitalizing letters of software names in the references
