+++ docs/refs.bib
@@ -335,7 +335,7 @@
 }
 
 @inproceedings{taghawi-nejad2017,
-  title = {ABCE: A Python Library for Economic Agent-Based Modeling},
+  title = {{ABCE}: A Python Library for Economic Agent-Based Modeling},
   shorttitle = {ABCE},
   booktitle = {Social Informatics},
   author = {{Taghawi-Nejad}, Davoud and Tanin, Rudy H. and Del Rio Chanona, R. Maria and Carro, Adri{\'a}n and Farmer, J. Doyne and Heinrich, Torsten and Sabuco, Juan and Straka, Mika J.},
