### 2024-03-29 ae4c83d
fix(docs): :pencil2: fixing capitalizing letters of software names in the references
