+++ docs/refs.bib
@@ -182,7 +182,7 @@
 }
 
 @article{foramitti2021,
-  title = {AgentPy: A package for agent-based modeling in Python},
+  title = {{AgentPy}: A package for agent-based modeling in {Python}},
   shorttitle = {AgentPy},
   author = {Foramitti, Jo{\"e}l},
   year = {2021},
@@ -233,8 +233,8 @@
 }
 
 @inproceedings{kazil2020a,
-  title = {Utilizing Python for Agent-Based Modeling: The {Mesa} Framework},
-  shorttitle = {Utilizing Python for Agent-Based Modeling},
+  title = {Utilizing {Python} for Agent-Based Modeling: The {Mesa} Framework},
+  shorttitle = {Utilizing {Python} for Agent-Based Modeling},
   booktitle = {Social, Cultural, and Behavioral Modeling},
   author = {Kazil, Jackie and Masad, David and Crooks, Andrew},
   editor = {Thomson, Robert and Bisgin, Halil and Dancy, Christopher and Hyder, Ayaz and Hussain, Muhammad},
@@ -335,7 +335,7 @@
 }
 
 @inproceedings{taghawi-nejad2017,
-  title = {{ABCE}: A Python Library for Economic Agent-Based Modeling},
+  title = {{ABCE}: A {Python} Library for Economic Agent-Based Modeling},
   shorttitle = {ABCE},
   booktitle = {Social Informatics},
   author = {{Taghawi-Nejad}, Davoud and Tanin, Rudy H. and Del Rio Chanona, R. Maria and Carro, Adri{\'a}n and Farmer, J. Doyne and Heinrich, Torsten and Sabuco, Juan and Straka, Mika J.},
@@ -372,7 +372,7 @@
 }
 
 @inproceedings{wang_mesa-geo_2022,
-  title = {{Mesa-Geo}: A GIS extension for the {Mesa} agent-based modeling framework in python},
+  title = {{Mesa-Geo}: A GIS extension for the {Mesa} agent-based modeling framework in {Python}},
   booktitle = {Proceedings of the 5th ACM SIGSPATIAL international workshop on GeoSpatial simulation},
   author = {Wang, Boyu and Hess, Vincent and Crooks, Andrew},
   year = {2022},
@@ -431,7 +431,7 @@
 }
 
 @article{hoyer2017xarray,
-  title     = {xarray: {N-D} labeled arrays and datasets in {Python}},
+  title     = {{Xarray}: {N-D} labeled arrays and datasets in {Python}},
   author    = {Hoyer, S. and J. Hamman},
   journal   = {Journal of Open Research Software},
   volume    = {5},
