### 2024-03-30 34493ed
Updated testing of the fire spread model
