### 2024-03-31 2efbf11
fix(src): :bug: np.nan is converted to True when masking array
