### 2024-03-31 80b9612
Added test for the hotelling law and fire examples
