### 2024-04-05 9df8b98
fix(docs): :pencil2: fixing typos according to JOSS editor's comments
