+++ docs/refs.bib
@@ -224,29 +224,31 @@
 }
 
 @article{janssen2008,
-  title = {Towards a community framework for agent-based modelling},
-  author = {Janssen, Marco and Alessa, Lilian and Barton, C Michael and Bergin, Sean and Lee, Allen},
-  year = {2008},
-  month = mar,
-  journal = {Journal of Artificial Societies and Social Simulation, The},
-  volume = {11}
+   title = {Towards a Community Framework for Agent-Based Modelling},
+   author = {Janssen, Marco A. and Alessa, Lilian Na'ia and Barton, Michael and Bergin, Sean and Lee, Allen},
+   journal = {Journal of Artificial Societies and Social Simulation},
+   ISSN = {1460-7425},
+   volume = {11},
+   number = {2},
+   pages = {6},
+   year = {2008},
+   URL = {https://www.jasss.org/11/2/6.html},
+   keywords = {Replication, Documentation Protocol, Software Development, Standardization, Test Beds, Education, Primitives},
+   abstract = {Agent-based modelling has become an increasingly important tool for scholars studying social and social-ecological systems, but there are no community standards on describing, implementing, testing and teaching these tools. This paper reports on the establishment of the Open Agent-Based Modelling Consortium, www.openabm.org, a community effort to foster the agent-based modelling development, communication, and dissemination for research, practice and education.},
 }
 
 @inproceedings{kazil2020a,
-  title = {Utilizing {Python} for Agent-Based Modeling: The {Mesa} Framework},
-  shorttitle = {Utilizing {Python} for Agent-Based Modeling},
-  booktitle = {Social, Cultural, and Behavioral Modeling},
+  title = {Utilizing {Python} for agent-based modeling: The {Mesa} framework},
+  booktitle = {Social, cultural, and behavioral modeling},
   author = {Kazil, Jackie and Masad, David and Crooks, Andrew},
   editor = {Thomson, Robert and Bisgin, Halil and Dancy, Christopher and Hyder, Ayaz and Hussain, Muhammad},
   year = {2020},
-  series = {Lecture Notes in Computer Science},
   pages = {308--317},
   publisher = {Springer International Publishing},
   address = {Cham},
-  doi = {10.1007/978-3-030-61255-9_30},
-  isbn = {978-3-030-61255-9},
-  language = {en}
+  isbn = {978-3-030-61255-9}
 }
+
 
 @misc{langevin2021modflow,
   title = {MODFLOW 6 modular hydrologic model version 6.2. 1: US geological survey software release, 18 february 2021 {{https://doi.}} org/10.5066},
@@ -335,20 +337,17 @@
 }
 
 @inproceedings{taghawi-nejad2017,
-  title = {{ABCE}: A {Python} Library for Economic Agent-Based Modeling},
-  shorttitle = {ABCE},
-  booktitle = {Social Informatics},
+  title = {{ABCE}: A {Python} library for economic agent-based modeling},
+  booktitle = {Social informatics},
   author = {{Taghawi-Nejad}, Davoud and Tanin, Rudy H. and Del Rio Chanona, R. Maria and Carro, Adri{\'a}n and Farmer, J. Doyne and Heinrich, Torsten and Sabuco, Juan and Straka, Mika J.},
   editor = {Ciampaglia, Giovanni Luca and Mashhadi, Afra and Yasseri, Taha},
   year = {2017},
-  series = {Lecture Notes in Computer Science},
   pages = {17--30},
   publisher = {Springer International Publishing},
   address = {Cham},
-  doi = {10.1007/978-3-319-67217-5_2},
-  isbn = {978-3-319-67217-5},
-  language = {en}
+  isbn = {978-3-319-67217-5}
 }
+
 
[…]
