### 2024-04-05 1a49092
fix(docs): :art: improve citation and figs formats
