### 2024-04-05 fc6ea75
fix(docs): :art: fig auto ref formats
