### 2024-04-05 c1589c4
fix(docs): :bug: try to fix citation formats problem
