+++ docs/refs.bib
@@ -237,7 +237,7 @@
    abstract = {Agent-based modelling has become an increasingly important tool for scholars studying social and social-ecological systems, but there are no community standards on describing, implementing, testing and teaching these tools. This paper reports on the establishment of the Open Agent-Based Modelling Consortium, www.openabm.org, a community effort to foster the agent-based modelling development, communication, and dissemination for research, practice and education.},
 }
 
-@inproceedings{kazil2020a,
+@book{kazil2020a,
   title = {Utilizing {Python} for agent-based modeling: The {Mesa} framework},
   booktitle = {Social, cultural, and behavioral modeling},
   author = {Kazil, Jackie and Masad, David and Crooks, Andrew},
@@ -336,7 +336,7 @@
   doi = {10.18564/jasss.3423}
 }
 
-@inproceedings{taghawi-nejad2017,
+@book{taghawi-nejad2017,
   title = {{ABCE}: A {Python} library for economic agent-based modeling},
   booktitle = {Social informatics},
   author = {{Taghawi-Nejad}, Davoud and Tanin, Rudy H. and Del Rio Chanona, R. Maria and Carro, Adri{\'a}n and Farmer, J. Doyne and Heinrich, Torsten and Sabuco, Juan and Straka, Mika J.},
