### 2024-04-05 #6298 @SongshGeo
@crvernon 👋 Hi,
I followed all the comments. Please check the newest version throughout.

I just wanted to let you know about the last two comments on citation formats. I confirmed that I used the `.bib` provided by the sources. However, their default category was `@InProceedding`. When I used this category, the colon didn't appear. I suppose this category is [not available in JOSS's document](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).  Therefore, I changed the categories to '@book'. Does it work?
