### 2024-04-05 #6298 @crvernon
👋  @SongshGeo - Thanks for making the requested changes to the paper! Next is just setting up the archive for your new release.

So here is what we have left to do:

- [ ]  Please respond with the DOI of the archived version here
