### 2024-04-05 80efbe8
fix(project): :art: fixing type hints througout
