+++ tests/api/test_modules.py
@@ -82,10 +82,10 @@ class TestCompositeModule:
     def test_create_non_module_raises_error(self, com_module: CompositeModule):
         """不是模块的类添加时会报错"""
         # arrange / act / assert
-        with pytest.raises(TypeError):
+        with pytest.raises(AssertionError):
             com_module.create_module(object)
 
-    def test_initialization(self, com_module: CompositeModule):
+    def test_initialization(self, com_module: CompositeModule) -> None:
         """测试初始化"""
         # arrange / act
         module_1 = com_module.create_module(Module, name="test_1")
