+++ abses/links.py
@@ -14,8 +14,11 @@ Actor, PatchCell can be used to create links.
 from __future__ import annotations
 
 import contextlib
+from abc import abstractmethod
+from functools import cached_property
 from typing import (
     TYPE_CHECKING,
+    Any,
     Dict,
     Iterable,
     List,
@@ -23,6 +26,8 @@ from typing import (
     Optional,
     Set,
     Tuple,
+    Union,
+    cast,
 )
 
 with contextlib.suppress(ImportError):
@@ -42,6 +47,9 @@ if TYPE_CHECKING:
 
 LinkingNode: TypeAlias = "Actor | PatchCell"
 Direction: TypeAlias = Optional[Literal["in", "out"]]
+DEFAULT_TARGETS: Tuple[str, str] = ("cell", "actor")
+TargetName: TypeAlias = Union[Literal["cell", "actor"], str]
+AttrGetter: TypeAlias = Union["Actor", "PatchCell", ActorsList["Actor"]]
 
 
 class _LinkContainer:
@@ -53,19 +61,18 @@ class _LinkContainer:
         self._cached_networks: Dict[str, object] = {}
 
     @property
-    def links(self) -> tuple[str]:
+    def links(self) -> Tuple[str, ...]:
         """Get the links of a certain type."""
         return tuple(self._links.keys())
 
-    @links.setter
-    def links(self, link_name: str) -> None:
-        """Set the links."""
+    def _add_a_link_name(self, link_name: str) -> None:
+        """Add a link."""
         self._links[link_name] = {}
         self._back_links[link_name] = {}
 
     def owns_links(
         self, node: LinkingNode, direction: Direction = "out"
-    ) -> Tuple[str]:
+    ) -> Tuple[str, ...]:
         """The links a specific node owns."""
         if direction == "out":
             data = self._links
@@ -104,7 +111,7 @@ class _LinkContainer:
     ) -> None:
         """Register a link."""
         if link_name not in self._links:
-            self.links = link_name
+            self._add_a_link_name(link_name)
         if source not in self._links[link_name]:
             self._links[link_name][source] = set()
         if target not in self._back_links[link_name]:
@@ -112,7 +119,7 @@ class _LinkContainer:
 
     def has_link(
         self, link_name: str, source: LinkingNode, target: LinkingNode
-    ) -> tuple[bool]:
+    ) -> Tuple[bool, bool]:
         """If any link exists between source and target.
 
         Parameters:
@@ -209,7 +216,7 @@ class _LinkContainer:
             link_name = [link_name]
         if not isinstance(link_name, Iterable):
             raise TypeError(f"{link_name} is not an iterable.")
-        return link_name
+        return list(link_name)
 
     def clean_links_of(
         self,
@@ -258,7 +265,7 @@ class _LinkContainer:
         link_name: Optional[str] = None,
         direction: Direction = None,
         default: bool = False,
-    ) -> ActorsList[LinkingNode]:
+    ) -> Set[LinkingNode]:
         """Get the linked nodes.
 
         Parameters:
@@ -283,12 +290,12 @@ class _LinkContainer:
         elif direction == "out":
             data = self._links
         elif direction is None:
-            return self.linked(node, link_name, direction="in") | self.linked(
-                node, link_name, direction="out"
-            )
+            in_links = self.linked(node, link_name, direction="in")
+            out_links = self.linked(node, link_name, direction="out")
+            return in_links | out_links
         else:
             raise ValueError(f"Invalid direction {direction}")
-        agents = set()
+        agents: Set[LinkingNode] = set()
         for name in link_names:
             if name not in data and default:
                 continue
@@ -300,7 +307,7 @@ class _LinkProxy:
[…]

[…]
