+++ README.md
@@ -3,6 +3,10 @@
 ![ABSES_banner](https://songshgeo-picgo-1302043007.cos.ap-beijing.myqcloud.com/uPic/CleanShot%202023-10-19%20at%2019.08.12@2x.png)
 
 <div align="center"><p>
+    <!-- JOSS badge -->
+    <!-- <a href="https://joss.theoj.org/papers/b6f6440bc05648fa7945bec1ae901e52">
+        <img alt="Latest release" src="https://joss.theoj.org/papers/b6f6440bc05648fa7945bec1ae901e52/status.svg?style=for-the-badge&logo=probot&color=C9CBFF&logoColor=D9E0EE&labelColor=302D41" />
+    </a> -->
     <a href="https://github.com/absespy/ABSESpy/releases/latest">
         <img alt="Latest release" src="https://img.shields.io/github/v/release/absespy/ABSESpy?style=for-the-badge&logo=probot&color=C9CBFF&logoColor=D9E0EE&labelColor=302D41" />
     </a>
@@ -58,7 +62,7 @@ Language: [English Readme] | [简体中文]
 - 📁 Handling complex models with a configuration file.
 - 🙋 and more...
 
-## Thanks to all contributors
+## Contributors
 
 <!-- ALL-CONTRIBUTORS-LIST:START - Do not remove or modify this section -->
 <!-- prettier-ignore-start -->
