### 2024-04-06 348799e
fix(docs): :label: fixing redundant title of README
