### 2024-04-06 #6298 @crvernon
:wave: @SongshGeo - Please change your declared license in the Zenodo archive (currently creative commons attribution 4.0 international) to match what you use for your software (Apache License 2.0).  Let me know when you have this done and I'll move forward.  Thanks!
