### 2024-04-06 52fa7b8
chore(project): :bookmark: update to 0.5.6 bookmark
