### 2024-04-08 a2c14c3
refactor(src): :recycle: creating agents from geopandas in AgentsContainer now
