### 2024-04-08 44d3f2d
refactor(tests): :recycle: seperate values getting setting
