### 2024-04-08 d876871
fix(src): :bug: AttributeError when getting value with wrong key
