### 2024-04-08 2818840
fix(src): :bug: setting values only when it has attribute
