### 2024-04-09 #6298 @SongshGeo
@crvernon  Hi, I cannot download the proof article from the above link. When I tried to check the PDF file from JOSS's GitHub repo, I got the following error:

> Error rendering embedded code
> Invalid PDF
