### 2024-04-09 #6298 @kthyng
@SongshGeo  Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. I think "earth" and "anthropocene" should be capitalized right?
