### 2024-04-10 6d33415
fix(docs): :memo: fixing Mantilla Ibarra name and capitalizing refers
