### 2024-04-10 #6298 @SongshGeo
@kthyng @crvernon @jamesdamillington @wang-boyu

Thanks for the efforts from you all. The comments helped us improve our package.
Sure! I'm delighted to help with potential reviews in the future. I signed up just now.

By the way, does anyone know how to change JOSS's badge style? It seems not supported by [Shields](https://shields.io/). I cannot add `?style=` to make it consistent with my other badges.
