### 2024-04-15 #6298 @kthyng
@SongshGeo I got an answer for you: We (JOSS) make them badges ourselves and they can't be re-styled. Sorry!
